@@ -26,6 +26,7 @@ class DashboardController extends Controller
     {
         DB::statement("SET sql_mode=(SELECT REPLACE(@@sql_mode,'ONLY_FULL_GROUP_BY',''));");
     }
+
     public function user_dashboard(Request $request)
     {
         $params = [
@@ -47,52 +48,52 @@ class DashboardController extends Controller
         $delivery_man = DeliveryMan::with('last_location')->when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })
-        ->Zonewise()
-        ->limit(2)->get('image');
+            ->Zonewise()
+            ->limit(2)->get('image');
 
         $active_deliveryman = DeliveryMan::when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })
-        ->Zonewise()->Active()->count();
+            ->Zonewise()->Active()->count();
 
         $inactive_deliveryman = DeliveryMan::when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })
-        ->Zonewise()->where('application_status','approved')->where('active',0)->count();
+            ->Zonewise()->where('application_status', 'approved')->where('active', 0)->count();
 
         $blocked_deliveryman = DeliveryMan::when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })
-        ->Zonewise()->where('application_status','approved')->where('status',0)->count();
+            ->Zonewise()->where('application_status', 'approved')->where('status', 0)->count();
 
         $newly_joined_deliveryman = DeliveryMan::when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
-        })
-        ->Zonewise()->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'))->count();
+        })->where('application_status', 'approved')
+            ->Zonewise()->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'))->count();
 
         $reviews = Review::when(is_numeric($params['zone_id']), function ($q) use ($params) {
-            return $q->whereHas('item.store', function ($query) use ($params){
+            return $q->whereHas('item.store', function ($query) use ($params) {
                 return $query->where('zone_id', $params['zone_id']);
             });
         })->count();
 
         $positive_reviews = Review::when(is_numeric($params['zone_id']), function ($q) use ($params) {
-            return $q->whereHas('item.store', function ($query) use ($params){
+            return $q->whereHas('item.store', function ($query) use ($params) {
                 return $query->where('zone_id', $params['zone_id']);
             });
-        })->whereIn('rating', [4,5])->get()->count();
+        })->whereIn('rating', [4, 5])->get()->count();
         $good_reviews = Review::when(is_numeric($params['zone_id']), function ($q) use ($params) {
-            return $q->whereHas('item.store', function ($query) use ($params){
+            return $q->whereHas('item.store', function ($query) use ($params) {
                 return $query->where('zone_id', $params['zone_id']);
             });
         })->where('rating', 3)->count();
         $neutral_reviews = Review::when(is_numeric($params['zone_id']), function ($q) use ($params) {
-            return $q->whereHas('item.store', function ($query) use ($params){
+            return $q->whereHas('item.store', function ($query) use ($params) {
                 return $query->where('zone_id', $params['zone_id']);
             });
         })->where('rating', 2)->count();
         $negative_reviews = Review::when(is_numeric($params['zone_id']), function ($q) use ($params) {
-            return $q->whereHas('item.store', function ($query) use ($params){
+            return $q->whereHas('item.store', function ($query) use ($params) {
                 return $query->where('zone_id', $params['zone_id']);
             });
         })->where('rating', 1)->count();
@@ -106,7 +107,7 @@ class DashboardController extends Controller
 
         $last_year_users = User::zone($params['zone_id'])
             ->whereMonth('created_at', 12)
-            ->whereYear('created_at', now()->format('Y')-1)
+            ->whereYear('created_at', now()->format('Y') - 1)
             ->count();
 
         $users = User::zone($params['zone_id'])
@@ -126,15 +127,15 @@ class DashboardController extends Controller
             }
         }
 
-        $active_customers = User::zone($params['zone_id'])->where('status',1)->count();
-        $blocked_customers = User::zone($params['zone_id'])->where('status',0)->count();
+        $active_customers = User::zone($params['zone_id'])->where('status', 1)->count();
+        $blocked_customers = User::zone($params['zone_id'])->where('status', 0)->count();
         $newly_joined = User::zone($params['zone_id'])->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'))->count();
 
-        $employees = Admin::zone()->with(['role'])->where('role_id', '!=','1')
-        ->when(is_numeric($params['zone_id']), function ($q) use ($params) {
-            return $q->where('zone_id', $params['zone_id']);
-        })
-        ->get();
+        $employees = Admin::zone()->with(['role'])->where('role_id', '!=', '1')
+            ->when(is_numeric($params['zone_id']), function ($q) use ($params) {
+                return $q->where('zone_id', $params['zone_id']);
+            })
+            ->get();
 
         $deliveryMen = DeliveryMan::with('last_location')->when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
@@ -143,7 +144,7 @@ class DashboardController extends Controller
         $deliveryMen = Helpers::deliverymen_list_formatting($deliveryMen);
 
         $module_type = Config::get('module.current_module_type');
-        return view("admin-views.dashboard-{$module_type}", compact('data','reviews','this_month','user_data','neutral_reviews','good_reviews','negative_reviews','positive_reviews','employees','active_deliveryman','deliveryMen','inactive_deliveryman','newly_joined_deliveryman','delivery_man', 'total_sell', 'commission', 'delivery_commission', 'params','module_type', 'customers','active_customers','blocked_customers', 'newly_joined','last_year_users', 'blocked_deliveryman'));
+        return view("admin-views.dashboard-{$module_type}", compact('data', 'reviews', 'this_month', 'user_data', 'neutral_reviews', 'good_reviews', 'negative_reviews', 'positive_reviews', 'employees', 'active_deliveryman', 'deliveryMen', 'inactive_deliveryman', 'newly_joined_deliveryman', 'delivery_man', 'total_sell', 'commission', 'delivery_commission', 'params', 'module_type', 'customers', 'active_customers', 'blocked_customers', 'newly_joined', 'last_year_users', 'blocked_deliveryman'));
     }
 
     public function transaction_dashboard(Request $request)
@@ -174,39 +175,38 @@ class DashboardController extends Controller
         $delivery_man = DeliveryMan::with('last_location')->when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })
-        ->Zonewise()
-        ->limit(2)->get('image');
+            ->Zonewise()
+            ->limit(2)->get('image');
 
         $active_deliveryman = DeliveryMan::when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })
-        ->Zonewise()->where('active',1)->count();
+            ->Zonewise()->where('active', 1)->count();
 
         $inactive_deliveryman = DeliveryMan::when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })
-        ->Zonewise()->where('application_status','approved')->where('active',0)->count();
+            ->Zonewise()->where('application_status', 'approved')->where('active', 0)->count();
 
         $suspend_deliveryman = DeliveryMan::when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })
-        ->Zonewise()->where('application_status','approved')->where('status',0)->count();
+            ->Zonewise()->where('application_status', 'approved')->where('status', 0)->count();
 
         $unavailable_deliveryman = DeliveryMan::when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })
-        ->Zonewise()->where('active',1)->Unavailable()->count();
+            ->Zonewise()->where('active', 1)->Unavailable()->count();
 
         $available_deliveryman = DeliveryMan::when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })
-        ->Zonewise()->where('active',1)->Available()->count();
+            ->Zonewise()->where('active', 1)->Available()->count();
 
         $newly_joined_deliveryman = DeliveryMan::when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })
-        ->Zonewise()->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'))->count();
-
+            ->Zonewise()->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'))->count();
         $deliveryMen = DeliveryMan::when(is_numeric($params['zone_id']), function ($q) use ($params) {
             return $q->where('zone_id', $params['zone_id']);
         })->zonewise()->available()->active()->get();
@@ -214,7 +214,7 @@ class DashboardController extends Controller
         $deliveryMen = Helpers::deliverymen_list_formatting($deliveryMen);
 
         $module_type = Config::get('module.current_module_type');
-        return view("admin-views.dashboard-{$module_type}", compact('data','active_deliveryman','deliveryMen','unavailable_deliveryman','available_deliveryman','inactive_deliveryman','newly_joined_deliveryman','delivery_man', 'total_sell', 'commission', 'delivery_commission','label', 'params','module_type','suspend_deliveryman'));
+        return view("admin-views.dashboard-{$module_type}", compact('data', 'active_deliveryman', 'deliveryMen', 'unavailable_deliveryman', 'available_deliveryman', 'inactive_deliveryman', 'newly_joined_deliveryman', 'delivery_man', 'total_sell', 'commission', 'delivery_commission', 'label', 'params', 'module_type', 'suspend_deliveryman'));
     }
 
     public function dashboard(Request $request)
@@ -234,16 +234,16 @@ class DashboardController extends Controller
         $delivery_commission = $data['delivery_commission'];
         $label = $data['label'];
         $module_type = Config::get('module.current_module_type');
-        if($module_type == 'settings'){
+        if ($module_type == 'settings') {
             return redirect()->route('admin.business-settings.business-setup');
         }
-        if($module_type == 'rental' && addon_published_status('Rental') == 1){
+        if ($module_type == 'rental' && addon_published_status('Rental') == 1) {
             return redirect()->route('admin.rental.dashboard');
         }
-        if($module_type == 'rental' && addon_published_status('Rental') == 0){
+        if ($module_type == 'rental' && addon_published_status('Rental') == 0) {
             return view('errors.404');
         }
-        return view("admin-views.dashboard-{$module_type}", compact('data', 'total_sell', 'commission', 'delivery_commission', 'label','params','module_type'));
+        return view("admin-views.dashboard-{$module_type}", compact('data', 'total_sell', 'commission', 'delivery_commission', 'label', 'params', 'module_type'));
 
     }
 
@@ -268,7 +268,7 @@ class DashboardController extends Controller
             return response()->json([
                 'view' => view('admin-views.partials._dashboard-order-stats-parcel', compact('data'))->render()
             ], 200);
-        }elseif($module_type == 'food'){
+        } elseif ($module_type == 'food') {
             return response()->json([
                 'view' => view('admin-views.partials._dashboard-order-stats-food', compact('data'))->render()
             ], 200);
@@ -309,14 +309,12 @@ class DashboardController extends Controller
             'top_selling_foods' => view('admin-views.partials._top-selling-foods', compact('top_sell'))->render(),
 
 
-
-
             'user_overview' => view('admin-views.partials._user-overview-chart', compact('data'))->render(),
             'monthly_graph' => view('admin-views.partials._monthly-earning-graph', compact('total_sell', 'commission', 'delivery_commission'))->render(),
             'stat_zone' => view('admin-views.partials._zone-change', compact('data'))->render(),
-            'order_stats' =>$module_type == 'parcel'? view('admin-views.partials._dashboard-order-stats-parcel', compact('data'))->render():
-            ($module_type == 'food'? view('admin-views.partials._dashboard-order-stats-food', compact('data'))->render():
-            view('admin-views.partials._dashboard-order-stats', compact('data'))->render()),
+            'order_stats' => $module_type == 'parcel' ? view('admin-views.partials._dashboard-order-stats-parcel', compact('data'))->render() :
+                ($module_type == 'food' ? view('admin-views.partials._dashboard-order-stats-food', compact('data'))->render() :
+                    view('admin-views.partials._dashboard-order-stats', compact('data'))->render()),
         ], 200);
     }
 
@@ -342,6 +340,7 @@ class DashboardController extends Controller
             'view' => view('admin-views.partials._user-overview-chart', compact('data'))->render()
         ], 200);
     }
+
     public function commission_overview(Request $request)
     {
         $params = session('dash_params');
@@ -375,18 +374,18 @@ class DashboardController extends Controller
             $refund_requested = Order::where('module_id', $module_id)->where(['order_status' => 'refund_requested'])->whereDate('refund_requested', Carbon::now());
             $refunded = Order::where('module_id', $module_id)->where(['order_status' => 'refunded'])->whereDate('refunded', Carbon::now());
             $new_orders = Order::where('module_id', $module_id)->whereDate('schedule_at', Carbon::now());
-            $new_items = Item::where('is_approved',1)->where('module_id', $module_id)->whereDate('created_at', Carbon::now());
-            $new_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id)->whereDate('created_at', Carbon::now());
+            $new_items = Item::where('is_approved', 1)->where('module_id', $module_id)->whereDate('created_at', Carbon::now());
+            $new_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id)->whereDate('created_at', Carbon::now());
             $new_customers = User::whereDate('created_at', Carbon::now());
-            if($module_type =='parcel'){
+            if ($module_type == 'parcel') {
                 $total_orders = Order::where('module_id', $module_id)->whereDate('created_at', Carbon::now());
-            } else{
+            } else {
                 $total_orders = Order::where('module_id', $module_id);
             }
-            $total_items = Item::where('is_approved',1)->where('module_id', $module_id);
-            $total_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id);
+            $total_items = Item::where('is_approved', 1)->where('module_id', $module_id);
+            $total_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id);
             $total_customers = User::all();
-        } elseif($module_id && $params['statistics_type'] == 'this_year'){
+        } elseif ($module_id && $params['statistics_type'] == 'this_year') {
             $searching_for_dm = Order::SearchingForDeliveryman()->where('module_id', $module_id)->whereYear('created_at', now()->format('Y'));
             $accepted_by_dm = Order::AccepteByDeliveryman()->where('module_id', $module_id)->whereYear('accepted', now()->format('Y'));
             $preparing_in_rs = Order::Preparing()->where('module_id', $module_id)->whereYear('processing', now()->format('Y'));
@@ -396,14 +395,14 @@ class DashboardController extends Controller
             $refund_requested = Order::where('module_id', $module_id)->where(['order_status' => 'refund_requested'])->whereYear('refund_requested', now()->format('Y'));
             $refunded = Order::where('module_id', $module_id)->where(['order_status' => 'refunded'])->whereYear('refunded', now()->format('Y'));
             $new_orders = Order::where('module_id', $module_id)->whereYear('schedule_at', now()->format('Y'));
-            $new_items = Item::where('is_approved',1)->where('module_id', $module_id)->whereYear('created_at', now()->format('Y'));
-            $new_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id)->whereYear('created_at', now()->format('Y'));
+            $new_items = Item::where('is_approved', 1)->where('module_id', $module_id)->whereYear('created_at', now()->format('Y'));
+            $new_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id)->whereYear('created_at', now()->format('Y'));
             $new_customers = User::whereYear('created_at', now()->format('Y'));
             $total_orders = Order::where('module_id', $module_id);
-            $total_items = Item::where('is_approved',1)->where('module_id', $module_id);
-            $total_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id);
+            $total_items = Item::where('is_approved', 1)->where('module_id', $module_id);
+            $total_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id);
             $total_customers = User::all();
-        } elseif($module_id && $params['statistics_type'] == 'this_month'){
+        } elseif ($module_id && $params['statistics_type'] == 'this_month') {
             $searching_for_dm = Order::SearchingForDeliveryman()->where('module_id', $module_id)->whereMonth('created_at', now()->format('m'))->whereYear('created_at', now()->format('Y'));
             $accepted_by_dm = Order::AccepteByDeliveryman()->where('module_id', $module_id)->whereMonth('accepted', now()->format('m'))->whereYear('accepted', now()->format('Y'));
             $preparing_in_rs = Order::Preparing()->where('module_id', $module_id)->whereMonth('processing', now()->format('m'))->whereYear('processing', now()->format('Y'));
@@ -413,14 +412,14 @@ class DashboardController extends Controller
             $refund_requested = Order::where('module_id', $module_id)->where(['order_status' => 'refund_requested'])->whereMonth('refund_requested', now()->format('m'))->whereYear('refund_requested', now()->format('Y'));
             $refunded = Order::where('module_id', $module_id)->where(['order_status' => 'refunded'])->whereMonth('refunded', now()->format('m'))->whereYear('refunded', now()->format('Y'));
             $new_orders = Order::where('module_id', $module_id)->whereMonth('schedule_at', now()->format('m'))->whereYear('schedule_at', now()->format('Y'));
-            $new_items = Item::where('is_approved',1)->where('module_id', $module_id)->whereMonth('created_at', now()->format('m'))->whereYear('created_at', now()->format('Y'));
-            $new_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id)->whereMonth('created_at', now()->format('m'))->whereYear('created_at', now()->format('Y'));
+            $new_items = Item::where('is_approved', 1)->where('module_id', $module_id)->whereMonth('created_at', now()->format('m'))->whereYear('created_at', now()->format('Y'));
+            $new_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id)->whereMonth('created_at', now()->format('m'))->whereYear('created_at', now()->format('Y'));
             $new_customers = User::whereMonth('created_at', now()->format('m'))->whereYear('created_at', now()->format('Y'));
             $total_orders = Order::where('module_id', $module_id);
-            $total_items = Item::where('is_approved',1)->where('module_id', $module_id);
-            $total_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id);
+            $total_items = Item::where('is_approved', 1)->where('module_id', $module_id);
+            $total_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id);
             $total_customers = User::all();
-        } elseif($module_id && $params['statistics_type'] == 'this_week'){
+        } elseif ($module_id && $params['statistics_type'] == 'this_week') {
             $searching_for_dm = Order::SearchingForDeliveryman()->where('module_id', $module_id)->whereBetween('created_at', [now()->startOfWeek()->format('Y-m-d H:i:s'), now()->endOfWeek()->format('Y-m-d H:i:s')]);
             $accepted_by_dm = Order::AccepteByDeliveryman()->where('module_id', $module_id)->whereBetween('accepted', [now()->startOfWeek()->format('Y-m-d H:i:s'), now()->endOfWeek()->format('Y-m-d H:i:s')]);
             $preparing_in_rs = Order::Preparing()->where('module_id', $module_id)->whereBetween('processing', [now()->startOfWeek()->format('Y-m-d H:i:s'), now()->endOfWeek()->format('Y-m-d H:i:s')]);
@@ -430,14 +429,14 @@ class DashboardController extends Controller
             $refund_requested = Order::where('module_id', $module_id)->where(['order_status' => 'refund_requested'])->whereBetween('refund_requested', [now()->startOfWeek()->format('Y-m-d H:i:s'), now()->endOfWeek()->format('Y-m-d H:i:s')]);
             $refunded = Order::where('module_id', $module_id)->where(['order_status' => 'refunded'])->whereBetween('refunded', [now()->startOfWeek()->format('Y-m-d H:i:s'), now()->endOfWeek()->format('Y-m-d H:i:s')]);
             $new_orders = Order::where('module_id', $module_id)->whereBetween('schedule_at', [now()->startOfWeek()->format('Y-m-d H:i:s'), now()->endOfWeek()->format('Y-m-d H:i:s')]);
-            $new_items = Item::where('is_approved',1)->where('module_id', $module_id)->whereBetween('created_at', [now()->startOfWeek()->format('Y-m-d H:i:s'), now()->endOfWeek()->format('Y-m-d H:i:s')]);
-            $new_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id)->whereBetween('created_at', [now()->startOfWeek()->format('Y-m-d H:i:s'), now()->endOfWeek()->format('Y-m-d H:i:s')]);
+            $new_items = Item::where('is_approved', 1)->where('module_id', $module_id)->whereBetween('created_at', [now()->startOfWeek()->format('Y-m-d H:i:s'), now()->endOfWeek()->format('Y-m-d H:i:s')]);
+            $new_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id)->whereBetween('created_at', [now()->startOfWeek()->format('Y-m-d H:i:s'), now()->endOfWeek()->format('Y-m-d H:i:s')]);
             $new_customers = User::whereBetween('created_at', [now()->startOfWeek()->format('Y-m-d H:i:s'), now()->endOfWeek()->format('Y-m-d H:i:s')]);
             $total_orders = Order::where('module_id', $module_id);
-            $total_items = Item::where('is_approved',1)->where('module_id', $module_id);
-            $total_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id);
+            $total_items = Item::where('is_approved', 1)->where('module_id', $module_id);
+            $total_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id);
             $total_customers = User::all();
-        } elseif($module_id) {
+        } elseif ($module_id) {
             $searching_for_dm = Order::SearchingForDeliveryman()->where('module_id', $module_id);
             $accepted_by_dm = Order::AccepteByDeliveryman()->where('module_id', $module_id);
             $preparing_in_rs = Order::Preparing()->where('module_id', $module_id);
@@ -447,12 +446,12 @@ class DashboardController extends Controller
             $refund_requested = Order::failed()->where('module_id', $module_id);
             $refunded = Order::Refunded()->where('module_id', $module_id);
             $new_orders = Order::where('module_id', $module_id)->whereDate('schedule_at', '>=', now()->subDays(30)->format('Y-m-d'));
-            $new_items = Item::where('is_approved',1)->where('module_id', $module_id)->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'));
-            $new_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id)->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'));
+            $new_items = Item::where('is_approved', 1)->where('module_id', $module_id)->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'));
+            $new_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id)->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'));
             $new_customers = User::whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'));
             $total_orders = Order::where('module_id', $module_id);
-            $total_items = Item::where('is_approved',1)->where('module_id', $module_id);
-            $total_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id);
+            $total_items = Item::where('is_approved', 1)->where('module_id', $module_id);
+            $total_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id);
             $total_customers = User::all();
         } else {
             $searching_for_dm = Order::SearchingForDeliveryman();
@@ -464,16 +463,16 @@ class DashboardController extends Controller
             $refund_requested = Order::failed();
             $refunded = Order::Refunded();
             $new_orders = Order::whereDate('schedule_at', '>=', now()->subDays(30)->format('Y-m-d'));
-            $new_items = Item::where('is_approved',1)->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'));
-            $new_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'));
+            $new_items = Item::where('is_approved', 1)->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'));
+            $new_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'));
             $new_customers = User::whereDate('created_at', '>=', now()->subDays(30)->format('Y-m-d'));
             $total_orders = Order::all();
-            $total_items = Item::where('is_approved',1)->get();
-            $total_stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->get();
+            $total_items = Item::where('is_approved', 1)->get();
+            $total_stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->get();
             $total_customers = User::all();
         }
 
-        if (is_numeric($zone_id) && $module_id &&  !in_array($module_type ,['parcel']) ) {
+        if (is_numeric($zone_id) && $module_id && !in_array($module_type, ['parcel'])) {
             $searching_for_dm = $searching_for_dm->StoreOrder()->OrderScheduledIn(30)->where('zone_id', $zone_id)->count();
             $accepted_by_dm = $accepted_by_dm->StoreOrder()->where('zone_id', $zone_id)->count();
             $preparing_in_rs = $preparing_in_rs->StoreOrder()->where('zone_id', $zone_id)->count();
@@ -490,7 +489,7 @@ class DashboardController extends Controller
             $new_items = $new_items->count();
             $new_stores = $new_stores->where('zone_id', $zone_id)->count();
             $new_customers = $new_customers->count();
-        } elseif($module_id && $module_type!='parcel') {
+        } elseif ($module_id && $module_type != 'parcel') {
             $searching_for_dm = $searching_for_dm->StoreOrder()->OrderScheduledIn(30)->count();
             $accepted_by_dm = $accepted_by_dm->StoreOrder()->count();
             $preparing_in_rs = $preparing_in_rs->StoreOrder()->count();
@@ -507,7 +506,7 @@ class DashboardController extends Controller
             $new_items = $new_items->count();
             $new_stores = $new_stores->count();
             $new_customers = $new_customers->count();
-        } elseif(is_numeric($zone_id) && $module_id && $module_type =='parcel') {
+        } elseif (is_numeric($zone_id) && $module_id && $module_type == 'parcel') {
             $searching_for_dm = $searching_for_dm->ParcelOrder()->OrderScheduledIn(30)->where('zone_id', $zone_id)->count();
             $accepted_by_dm = $accepted_by_dm->ParcelOrder()->where('zone_id', $zone_id)->count();
             $preparing_in_rs = $preparing_in_rs->ParcelOrder()->where('zone_id', $zone_id)->count();
@@ -524,8 +523,7 @@ class DashboardController extends Controller
             $new_items = $new_items->count();
             $new_stores = $new_stores->where('zone_id', $zone_id)->count();
             $new_customers = $new_customers->where('zone_id', $zone_id)->count();
-        }
-        elseif($module_id && $module_type =='parcel') {
+        } elseif ($module_id && $module_type == 'parcel') {
             $searching_for_dm = $searching_for_dm->ParcelOrder()->OrderScheduledIn(30)->count();
             $accepted_by_dm = $accepted_by_dm->ParcelOrder()->count();
             $preparing_in_rs = $preparing_in_rs->ParcelOrder()->count();
@@ -542,9 +540,7 @@ class DashboardController extends Controller
             $new_items = $new_items->count();
             $new_stores = $new_stores->count();
             $new_customers = $new_customers->count();
-        }
-
-        else{
+        } else {
             $searching_for_dm = $searching_for_dm->StoreOrder()->OrderScheduledIn(30)->count();
             $accepted_by_dm = $accepted_by_dm->StoreOrder()->count();
             $preparing_in_rs = $preparing_in_rs->StoreOrder()->count();
@@ -590,11 +586,11 @@ class DashboardController extends Controller
         //zone
         if (is_numeric($zone_id)) {
             $customer = User::where('zone_id', $zone_id);
-            $stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id)->where(['zone_id' => $zone_id]);
+            $stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id)->where(['zone_id' => $zone_id]);
             $delivery_man = DeliveryMan::where('application_status', 'approved')->where('zone_id', $zone_id)->Zonewise();
         } else {
             $customer = User::whereNotNull('id');
-            $stores = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->where('module_id', $module_id)->whereNotNull('id');
+            $stores = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->where('module_id', $module_id)->whereNotNull('id');
             $delivery_man = DeliveryMan::where('application_status', 'approved')->Zonewise();
         }
         //user overview
@@ -602,14 +598,14 @@ class DashboardController extends Controller
             $customer = $customer->count();
             $stores = $stores->count();
             $delivery_man = $delivery_man->count();
-        } elseif($params['user_overview'] == 'this_month') {
+        } elseif ($params['user_overview'] == 'this_month') {
             $customer = $customer->whereMonth('created_at', date('m'))
                 ->whereYear('created_at', date('Y'))->count();
             $stores = $stores->whereMonth('created_at', date('m'))
                 ->whereYear('created_at', date('Y'))->count();
             $delivery_man = $delivery_man->whereMonth('created_at', date('m'))
                 ->whereYear('created_at', date('Y'))->count();
-        } elseif($params['user_overview'] == 'this_year') {
+        } elseif ($params['user_overview'] == 'this_year') {
             $customer = $customer
                 ->whereYear('created_at', date('Y'))->count();
             $stores = $stores
@@ -634,8 +630,8 @@ class DashboardController extends Controller
     {
         $params = session('dash_params');
         if (!url()->current() == $request->is('admin/users')) {
-        $data_os = self::order_stats_calc($params['zone_id'], $params['module_id']);
-        $data_uo = self::user_overview_calc($params['zone_id'], $params['module_id']);
+            $data_os = self::order_stats_calc($params['zone_id'], $params['module_id']);
+            $data_uo = self::user_overview_calc($params['zone_id'], $params['module_id']);
         }
         $popular = Wishlist::with(['store'])
             ->whereHas('store')
@@ -650,7 +646,7 @@ class DashboardController extends Controller
                 });
             })
             ->select('store_id', DB::raw('COUNT(store_id) as count'))->groupBy('store_id')
-            ->having("count" , '>', 0)
+            ->having("count", '>', 0)
             ->orderBy('count', 'DESC')
             ->limit(6)->get();
         $top_sell = Item::withoutGlobalScope(ZoneScope::class)
@@ -664,7 +660,7 @@ class DashboardController extends Controller
                     return $query->where('module_id', $params['module_id'])->where('zone_id', $params['zone_id']);
                 });
             })
-            ->having("order_count" , '>', 0)
+            ->having("order_count", '>', 0)
             ->orderBy("order_count", 'desc')
             ->take(6)
             ->get();
@@ -679,35 +675,35 @@ class DashboardController extends Controller
                     return $query->where('zone_id', $params['zone_id']);
                 });
             })
-            ->having("rating_count" , '>', 0)
+            ->having("rating_count", '>', 0)
             ->orderBy('rating_count', 'desc')
             ->take(6)
             ->get();
 
         $top_deliveryman = DeliveryMan::withCount('orders')->when(is_numeric($params['zone_id']), function ($q) use ($params) {
-                return $q->where('zone_id', $params['zone_id']);
-            })
+            return $q->where('zone_id', $params['zone_id']);
+        })
             ->Zonewise()
-            ->having("orders_count" , '>', 0)
+            ->having("orders_count", '>', 0)
             ->orderBy("orders_count", 'desc')
             ->take(6)
             ->get();
 
         $top_customers = User::when(is_numeric($params['zone_id']), function ($q) use ($params) {
-                return $q->where('zone_id', $params['zone_id']);
-            })
-            ->having("order_count" , '>', 0)
+            return $q->where('zone_id', $params['zone_id']);
+        })
+            ->having("order_count", '>', 0)
             ->orderBy("order_count", 'desc')
             ->take(6)
             ->get();
 
-        $top_restaurants = Store::whereHas('vendor',fn ($query) => $query->where('status',1))->when(is_numeric($params['module_id']), function ($q) use ($params) {
-                return $q->where('module_id', $params['module_id']);
-            })
+        $top_restaurants = Store::whereHas('vendor', fn($query) => $query->where('status', 1))->when(is_numeric($params['module_id']), function ($q) use ($params) {
+            return $q->where('module_id', $params['module_id']);
+        })
             ->when(is_numeric($params['zone_id']), function ($q) use ($params) {
                 return $q->where('zone_id', $params['zone_id']);
             })
-            ->having("order_count" , '>', 0)
+            ->having("order_count", '>', 0)
             ->orderBy("order_count", 'desc')
             ->take(6)
             ->get();
@@ -715,38 +711,38 @@ class DashboardController extends Controller
 
         // custom filtering for bar chart
         $months = array(
-            '"'.translate('Jan').'"',
-            '"'.translate('Feb').'"',
-            '"'.translate('Mar').'"',
-            '"'.translate('Apr').'"',
-            '"'.translate('May').'"',
-            '"'.translate('Jun').'"',
-            '"'.translate('Jul').'"',
-            '"'.translate('Aug').'"',
-            '"'.translate('Sep').'"',
-            '"'.translate('Oct').'"',
-            '"'.translate('Nov').'"',
-            '"'.translate('Dec').'"'
+            '"' . translate('Jan') . '"',
+            '"' . translate('Feb') . '"',
+            '"' . translate('Mar') . '"',
+            '"' . translate('Apr') . '"',
+            '"' . translate('May') . '"',
+            '"' . translate('Jun') . '"',
+            '"' . translate('Jul') . '"',
+            '"' . translate('Aug') . '"',
+            '"' . translate('Sep') . '"',
+            '"' . translate('Oct') . '"',
+            '"' . translate('Nov') . '"',
+            '"' . translate('Dec') . '"'
         );
         $days = array(
-            '"'.translate('Mon').'"',
-            '"'.translate('Tue').'"',
-            '"'.translate('Wed').'"',
-            '"'.translate('Thu').'"',
-            '"'.translate('Fri').'"',
-            '"'.translate('Sat').'"',
-            '"'.translate('Sun').'"',
+            '"' . translate('Mon') . '"',
+            '"' . translate('Tue') . '"',
+            '"' . translate('Wed') . '"',
+            '"' . translate('Thu') . '"',
+            '"' . translate('Fri') . '"',
+            '"' . translate('Sat') . '"',
+            '"' . translate('Sun') . '"',
         );
         $total_sell = [];
         $commission = [];
         $label = [];
         $query = OrderTransaction::NotRefunded()
-        ->when(is_numeric($params['module_id']), function ($q) use ($params) {
-            return $q->where('module_id', $params['module_id']);
-        })
-        ->when(is_numeric($params['zone_id']), function ($q) use ($params) {
-            return $q->where('zone_id', $params['zone_id']);
-        });
+            ->when(is_numeric($params['module_id']), function ($q) use ($params) {
+                return $q->where('module_id', $params['module_id']);
+            })
+            ->when(is_numeric($params['zone_id']), function ($q) use ($params) {
+                return $q->where('zone_id', $params['zone_id']);
+            });
         switch ($params['commission_overview']) {
             case "this_year":
                 for ($i = 1; $i <= 12; $i++) {
