@@ -338,13 +338,25 @@ class GenerateAdminRoute extends Command
             'admin-views.business-settings.landing-page-settings.admin-landing-download-apps' => ['admin/business-settings/pages/admin-landing-page-settings/download-apps'],
             'admin-views.business-settings.landing-page-settings.admin-landing-contact' => ['admin/business-settings/pages/admin-landing-page-settings/contact-us'],
             'admin-views.business-settings.landing-page-settings.admin-landing-background-color' => ['admin/business-settings/pages/admin-landing-page-settings/background-color'],
+
             'admin-views.business-settings.landing-page-settings.react-landing-page-header' => ['admin/business-settings/pages/react-landing-page-settings/header'],
+            'admin-views.business-settings.landing-page-settings.react-landing-page-trust-section' => ['admin/business-settings/pages/react-landing-page-settings/trust-section'],
+            'admin-views.business-settings.landing-page-settings.react-landing-available-zone' => ['admin/business-settings/pages/react-landing-page-settings/available-zone'],
+            'admin-views.business-settings.landing-page-settings.react-landing-promotion-banners' => ['admin/business-settings/pages/react-landing-page-settings/promotion-banner'],
+            'admin-views.business-settings.landing-page-settings.react-landing-download-apps' => ['admin/business-settings/pages/react-landing-page-settings/download-user-app'],
+            'admin-views.business-settings.landing-page-settings.react-landing-page-popular-clients' => ['admin/business-settings/pages/react-landing-page-settings/popular-clients'],
+            'admin-views.business-settings.landing-page-settings.react-landing-page-download-seller-app' => ['admin/business-settings/pages/react-landing-page-settings/download-seller-app'],
+            'admin-views.business-settings.landing-page-settings.react-landing-page-download-deliveryman-app' => ['admin/business-settings/pages/react-landing-page-settings/download-deliveryman-app'],
+            'admin-views.business-settings.landing-page-settings.react-landing-page-banner-section' => ['admin/business-settings/pages/react-landing-page-settings/banner-section'],
+            'admin-views.business-settings.landing-page-settings.react-landing-testimonial' => ['admin/business-settings/pages/react-landing-page-settings/testimonials'],
+            'admin-views.business-settings.landing-page-settings.react-landing-page-gallery' => ['admin/business-settings/pages/react-landing-page-settings/gallery'],
+            'admin-views.business-settings.landing-page-settings.react-landing-page-highlight-section' => ['admin/business-settings/pages/react-landing-page-settings/highlight-section'],
+            'admin-views.business-settings.landing-page-settings.react-landing-page-faq' => ['admin/business-settings/pages/react-landing-page-settings/faq'],
+            'admin-views.business-settings.landing-page-settings.react-landing-fixed-data' => ['admin/business-settings/pages/react-landing-page-settings/fixed-data'],
             'admin-views.business-settings.landing-page-settings.react-landing-page-company' => ['admin/business-settings/pages/react-landing-page-settings/company-intro'],
             'admin-views.business-settings.landing-page-settings.react-landing-earn-money' => ['admin/business-settings/pages/react-landing-page-settings/earn-money'],
-            'admin-views.business-settings.landing-page-settings.react-landing-available-zone' => ['admin/business-settings/pages/react-landing-page-settings/available-zone'],
             'admin-views.business-settings.landing-page-settings.react-landing-business' => ['admin/business-settings/pages/react-landing-page-settings/business-section'],
-            'admin-views.business-settings.landing-page-settings.react-landing-testimonial' => ['admin/business-settings/pages/react-landing-page-settings/testimonials'],
-            'admin-views.business-settings.landing-page-settings.react-landing-fixed-data' => ['admin/business-settings/pages/react-landing-page-settings/fixed-data'],
+
             'admin-views.business-settings.landing-page-settings.flutter-fixed-data' => ['admin/business-settings/pages/flutter-landing-page-settings/fixed-data'],
             'admin-views.business-settings.landing-page-settings.flutter-landing-page-special-criteria' => ['admin/business-settings/pages/flutter-landing-page-settings/special-criteria'],
             'admin-views.business-settings.landing-page-settings.flutter-landing-page-available-zone' => ['admin/business-settings/pages/flutter-landing-page-settings/available-zone'],
