@@ -66,26 +66,35 @@ class VendorController extends Controller
             'l_name' => 'nullable|max:100',
             'name.0' => 'required',
             'name.*' => 'max:191',
-            'address' => 'required|max:1000',
-            'latitude' => 'required',
-            'longitude' => 'required',
+            'address.0' => 'required',
+            'address.*' => 'max:1000',
+            'latitude' => 'required|numeric',
+            'longitude' => 'required|numeric',
             'email' => 'required|unique:vendors',
             'phone' => 'required|regex:/^([0-9\s\-\+\(\)]*)$/|min:10|max:20|unique:vendors',
             'minimum_delivery_time' => 'required',
             'maximum_delivery_time' => 'required',
             'delivery_time_type'=>'required',
-            'password' => ['required', Password::min(8)->mixedCase()->letters()->numbers()->symbols()->uncompromised(),
+            'password' => ['required', Password::min(8)->mixedCase()->letters()->numbers()->symbols(),
                 function ($attribute, $value, $fail) {
                     if (strpos($value, ' ') !== false) {
                         $fail('The :attribute cannot contain white spaces.');
                     }
                 },],
             'zone_id' => 'required',
-            'logo' => 'required',
+            'logo' => 'required|image|max:2048|mimes:'.IMAGE_FORMAT_FOR_VALIDATION,
+            'cover_photo' => 'nullable|image|max:2048|mimes:'.IMAGE_FORMAT_FOR_VALIDATION,
+
         ], [
             'f_name.required' => translate('messages.first_name_is_required'),
             'name.0.required'=>translate('default_name_is_required'),
+            'address.0.required'=>translate('default_address_is_required'),
         ]);
+
+
+      if ($validator->fails()) {
+           return response()->json(['errors' => Helpers::error_processor($validator)]);
+        }
 
         if($request->zone_id)
         {
@@ -95,23 +104,19 @@ class VendorController extends Controller
             ->first();
             if(!$zone){
                 $validator->getMessageBag()->add('latitude', translate('messages.coordinates_out_of_zone'));
-                return back()->withErrors($validator)
-                        ->withInput();
+            return response()->json(['errors' => Helpers::error_processor($validator)]);
             }
         }
+
         if ($request->delivery_time_type == 'min') {
             $minimum_delivery_time = (int) $request->input('minimum_delivery_time');
             if ($minimum_delivery_time < 10) {
                 $validator->getMessageBag()->add('minimum_delivery_time', translate('messages.minimum_delivery_time_should_be_more_than_10_min'));
-                return back()->withErrors($validator)
-                        ->withInput();
+            return response()->json(['errors' => Helpers::error_processor($validator)]);
             }
         }
-        if ($validator->fails()) {
-            return back()
-            ->withErrors($validator)
-            ->withInput();
-        }
+
+
 
         $vendor = new Vendor();
         $vendor->f_name = $request->f_name;
@@ -145,60 +150,16 @@ class VendorController extends Controller
             {
                 StoreLogic::insert_schedule($store->id);
             }
-            $default_lang = str_replace('_', '-', app()->getLocale());
-            $data = [];
-            foreach ($request->lang as $index => $key) {
-                if($default_lang == $key && !($request->name[$index])){
-                    if ($key != 'default') {
-                        array_push($data, array(
-                            'translationable_type' => 'App\Models\Store',
-                            'translationable_id' => $store->id,
-                            'locale' => $key,
-                            'key' => 'name',
-                            'value' => $store->name,
-                        ));
-                    }
-                }else{
-                    if ($request->name[$index] && $key != 'default') {
-                        array_push($data, array(
-                            'translationable_type' => 'App\Models\Store',
-                            'translationable_id' => $store->id,
-                            'locale' => $key,
-                            'key' => 'name',
-                            'value' => $request->name[$index],
-                        ));
-                    }
-                }
-                if($default_lang == $key && !($request->address[$index])){
-                    if ($key != 'default') {
-                        array_push($data, array(
-                            'translationable_type' => 'App\Models\Store',
-                            'translationable_id' => $store->id,
-                            'locale' => $key,
-                            'key' => 'address',
-                            'value' => $store->address,
-                        ));
-                    }
-                }else{
-                    if ($request->address[$index] && $key != 'default') {
-                        array_push($data, array(
-                            'translationable_type' => 'App\Models\Store',
-                            'translationable_id' => $store->id,
-                            'locale' => $key,
-                            'key' => 'address',
-                            'value' => $request->address[$index],
-                        ));
-                    }
-                }
-            }
-            Translation::insert($data);
-            // $store->zones()->attach($request->zone_ids);
-            //code...
+
+        Helpers::add_or_update_translations(request: $request, key_data: 'name', name_field: 'name', model_name: 'Store', data_id: $store->id, data_value: $store->name);
+        Helpers::add_or_update_translations(request: $request, key_data: 'address', name_field: 'address', model_name: 'Store', data_id: $store->id, data_value: $store->address);
+
         } catch (\Exception $ex) {
             info($ex->getMessage());
+            $validator->getMessageBag()->add('store_add', $ex->getMessage());
+            return response()->json(['errors' => Helpers::error_processor($validator)]);
         }
-        Toastr::success(translate('messages.store_added_successfully'));
-        return redirect('admin/store/list');
+      return response()->json(200);
     }
 
     public function edit($id)
@@ -217,23 +178,35 @@ class VendorController extends Controller
         $validator = Validator::make($request->all(), [
             'f_name' => 'required|max:100',
             'l_name' => 'nullable|max:100',
-            'name' => 'required|max:191',
+              'name.0' => 'required',
+            'name.*' => 'max:191',
+            'address.0' => 'required',
+            'address.*' => 'max:1000',
             'email' => 'required|unique:vendors,email,'.$store->vendor->id,
             'phone' => 'required|regex:/^([0-9\s\-\+\(\)]*)$/|min:10|max:20|unique:vendors,phone,'.$store->vendor->id,
             'zone_id'=>'required',
-            'latitude' => 'required',
-            'longitude' => 'required',
-            'password' => ['nullable', Password::min(8)->mixedCase()->letters()->numbers()->symbols()->uncompromised(),function ($attribute, $value, $fail) {
+            'latitude' => 'required|numeric',
+            'longitude' => 'required|numeric',
+            'password' => ['nullable', Password::min(8)->mixedCase()->letters()->numbers()->symbols(),function ($attribute, $value, $fail) {
                 if (strpos($value, ' ') !== false) {
                     $fail('The :attribute cannot contain white spaces.');
                 }
             },],
             'minimum_delivery_time' => 'required',
             'maximum_delivery_time' => 'required',
-            'delivery_time_type'=>'required'
+            'delivery_time_type'=>'required',
+            'logo' => 'nullable|image|max:2048|mimes:'.IMAGE_FORMAT_FOR_VALIDATION,
+            'cover_photo' => 'nullable|image|max:2048|mimes:'.IMAGE_FORMAT_FOR_VALIDATION,
         ], [
-            'f_name.required' => translate('messages.first_name_is_required')
+            'f_name.required' => translate('messages.first_name_is_required'),
+            'name.0.required'=>translate('default_name_is_required'),
+            'address.0.required'=>translate('default_address_is_required'),
         ]);
+
+           if ($validator->fails()) {
+           return response()->json(['errors' => Helpers::error_processor($validator)]);
+
+        }
 
         if($request->zone_id)
         {
@@ -243,24 +216,18 @@ class VendorController extends Controller
             ->first();
             if(!$zone){
                 $validator->getMessageBag()->add('latitude', translate('messages.coordinates_out_of_zone'));
-                return back()->withErrors($validator)
-                        ->withInput();
+                 return response()->json(['errors' => Helpers::error_processor($validator)]);
             }
         }
         if ($request->delivery_time_type == 'min') {
             $minimum_delivery_time = (int) $request->input('minimum_delivery_time');
             if ($minimum_delivery_time < 10) {
                 $validator->getMessageBag()->add('minimum_delivery_time', translate('messages.minimum_delivery_time_should_be_more_than_10_min'));
-                return back()->withErrors($validator)
-                        ->withInput();
+                return response()->json(['errors' => Helpers::error_processor($validator)]);
             }
         }
 
-        if ($validator->fails()) {
-            return back()
-                    ->withErrors($validator)
-                    ->withInput();
-        }
+
         $vendor = Vendor::findOrFail($store->vendor->id);
         $vendor->f_name = $request->f_name;
         $vendor->l_name = $request->l_name;
@@ -300,7 +267,6 @@ class VendorController extends Controller
         }
 
 
-        Toastr::success(translate('messages.store_updated_successfully'));
         if($request->approve_vendor == 1){
             $request->merge([
                 'status' => 1,
@@ -308,7 +274,8 @@ class VendorController extends Controller
             ]);
             $this->updateVendorApplication($request);
         }
-        return redirect('admin/store/list');
+
+        return response()->json(500);
     }
 
     public function destroy(Request $request, Store $store)
@@ -628,7 +595,7 @@ class VendorController extends Controller
         ->first();
 
         $total_transaction = $result->total_transaction;
-        $comission_earned = $result->commission_earned;
+        $comission_earned = max(0,$result->commission_earned);
 
         $store_withdraws = WithdrawRequest::wherehas('store', function($query){
             $query->where('module_id', Config::get('module.current_module_id'));
@@ -642,84 +609,59 @@ class VendorController extends Controller
 
     public function pending_requests(Request $request)
     {
+        $stores = $this->getNewStores($request, null);
+        $zone_id = $request->query('zone_id', 'all');
+        $type = $request->query('type', 'all');
+        $search_by = $request->query('search_by');
+         $zone = is_numeric($zone_id)?Zone::findOrFail($zone_id):null;
+        return view('admin-views.vendor.pending_requests', compact('stores', 'zone','type', 'search_by'));
+    }
+
+    private function getNewStores($request, $storeApproveStatus){
+
         $zone_id = $request->query('zone_id', 'all');
         $search_by = $request->query('search_by');
         $key = explode(' ', $search_by);
         $type = $request->query('type', 'all');
         $module_id = $request->query('module_id', 'all');
-        $stores = Store::with('vendor','module')->whereHas('vendor', function($query){
-            return $query->where('status', null);
-        })
-        ->when(is_numeric($zone_id), function($query)use($zone_id){
-                return $query->where('zone_id', $zone_id);
-        })
-        ->when(is_numeric($module_id), function($query)use($request){
-            return $query->module($request->query('module_id'));
-        })
-        ->when($search_by, function($query)use($key){
-            return $query->where(function($query)use($key){
-                $query->orWhereHas('vendor',function ($q) use ($key) {
-                    $q->where(function($q)use($key){
-                        foreach ($key as $value) {
-                            $q->orWhere('f_name', 'like', "%{$value}%")
-                                ->orWhere('l_name', 'like', "%{$value}%")
-                                ->orWhere('email', 'like', "%{$value}%")
-                                ->orWhere('phone', 'like', "%{$value}%");
-                        }
+
+
+        $stores = Store::with('vendor:id,f_name,l_name,status','module:id,module_name','zone:id,name')->whereHas('vendor', function($query) use($storeApproveStatus){
+                    return $query->where('status', $storeApproveStatus);
+                })
+                ->when(is_numeric($zone_id), function($query)use($zone_id){
+                        return $query->where('zone_id', $zone_id);
+                })
+                ->when(is_numeric($module_id), function($query)use($request){
+                    return $query->module($request->query('module_id'));
+                })
+                ->when($search_by, function($query)use($key){
+                    return $query->where(function($query)use($key){
+                        $query->orWhereHas('vendor',function ($q) use ($key) {
+                            $q->where(function($q)use($key){
+                                foreach ($key as $value) {
+                                    $q->orWhereAny([ 'f_name', 'l_name', 'email', 'phone' ], 'like', "%{$value}%");
+                                }
+                            });
+                        })->orWhere(function ($q) use ($key) {
+                            foreach ($key as $value) {
+                                $q->orWhereAny([  'name', 'email', 'phone'], 'like', "%{$value}%");
+                            }
+                        });
                     });
-                })->orWhere(function ($q) use ($key) {
-                    foreach ($key as $value) {
-                        $q->orWhere('name', 'like', "%{$value}%")
-                            ->orWhere('email', 'like', "%{$value}%")
-                            ->orWhere('phone', 'like', "%{$value}%");
-                    }
-                });
-            });
-        })
-        ->module(Config::get('module.current_module_id'))
-        ->type($type)->latest()->paginate(config('default_pagination'));
-        $zone = is_numeric($zone_id)?Zone::findOrFail($zone_id):null;
-        return view('admin-views.vendor.pending_requests', compact('stores', 'zone','type', 'search_by'));
+                })
+                ->module(Config::get('module.current_module_id'))
+                ->type($type)->latest()->paginate(config('default_pagination'));
+
+        return $stores;
     }
 
     public function deny_requests(Request $request)
     {
         $search_by = $request->query('search_by');
-        $key = explode(' ', $search_by);
         $zone_id = $request->query('zone_id', 'all');
         $type = $request->query('type', 'all');
-        $module_id = $request->query('module_id', 'all');
-        $stores = Store::with('vendor','module')->whereHas('vendor', function($query){
-            return $query->where('status', 0);
-        })
-        ->when(is_numeric($zone_id), function($query)use($zone_id){
-                return $query->where('zone_id', $zone_id);
-        })
-        ->when(is_numeric($module_id), function($query)use($request){
-            return $query->module($request->query('module_id'));
-        })
-        ->when($search_by, function($query)use($key){
-            return $query->where(function($query)use($key){
-                $query->orWhereHas('vendor',function ($q) use ($key) {
-                    $q->where(function($q)use($key){
-                        foreach ($key as $value) {
-                            $q->orWhere('f_name', 'like', "%{$value}%")
-                                ->orWhere('l_name', 'like', "%{$value}%")
-                                ->orWhere('email', 'like', "%{$value}%")
-                                ->orWhere('phone', 'like', "%{$value}%");
-                        }
-                    });
-                })->orWhere(function ($q) use ($key) {
-                    foreach ($key as $value) {
-                        $q->orWhere('name', 'like', "%{$value}%")
-                            ->orWhere('email', 'like', "%{$value}%")
-                            ->orWhere('phone', 'like', "%{$value}%");
-                    }
-                });
-            });
-        })
-        ->module(Config::get('module.current_module_id'))
-        ->type($type)->latest()->paginate(config('default_pagination'));
+        $stores = $this->getNewStores($request, 0);
         $zone = is_numeric($zone_id)?Zone::findOrFail($zone_id):null;
         return view('admin-views.vendor.deny_requests', compact('stores', 'zone','type', 'search_by'));
     }
@@ -790,10 +732,6 @@ class VendorController extends Controller
         $data = Store::whereHas('module', function($q)use($request){
             $q->whereNot('module_type', 'rental');
         })->
-        // withOutGlobalScopes()
-        // ->
-        // join('zones', 'zones.id', '=', 'stores.zone_id')
-        // ->
         when($zone_ids, function($query) use($zone_ids){
             $query->whereIn('stores.zone_id', [$zone_ids]);
         })
@@ -1100,7 +1038,7 @@ class VendorController extends Controller
                     $add_days=$store?->store_sub_update_application->validity;
                 }
                     $store?->store_sub_update_application->update([
-                        'expiry_date'=> Carbon::now()->addDays($add_days)->format('Y-m-d'),
+                        'expiry_date'=> Carbon::now()->addDays((int) $add_days)->format('Y-m-d'),
                         'status'=>1
                     ]);
                 $store->store_business_model= 'subscription';
