@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\api\v1\auth;
+namespace App\Http\Controllers\Api\V1\Auth;
 
 
 use App\CentralLogics\Helpers;
