@@ -2,82 +2,81 @@
 
 namespace App\Http\Controllers\Api\V1;
 
+ 
+
 ini_set('memory_limit', '-1');
 
-use App\Models\Order;
-use App\Library\Payer;
-use App\Models\OrderTransaction;
-use App\Traits\Payment;
-use App\Library\Receiver;
-use App\Models\DeliveryMan;
-use App\Models\Notification;
-use App\Models\OrderPayment;
-use Carbon\Carbon;
-use Illuminate\Http\Request;
 use App\CentralLogics\Helpers;
+use App\CentralLogics\OrderLogic;
+use App\Http\Controllers\Controller;
+use App\Library\Payer;
+use App\Library\Payment as PaymentInfo;
+use App\Library\Receiver;
+use App\Mail\WithdrawRequestMail;
+use App\Models\AccountTransaction;
+use App\Models\Admin;
 use App\Models\BusinessSetting;
 use App\Models\DeliveryHistory;
+use App\Models\DeliveryMan;
+use App\Models\DeliveryManWallet;
+use App\Models\DisbursementDetails;
+use App\Models\DisbursementWithdrawalMethod;
+use App\Models\Notification;
+use App\Models\Order;
+use App\Models\OrderPayment;
+use App\Models\OrderTransaction;
 use App\Models\ProvideDMEarning;
 use App\Models\UserNotification;
 use App\Models\WithdrawalMethod;
-use App\CentralLogics\OrderLogic;
-use App\Models\DeliveryManWallet;
-use App\Models\AccountTransaction;
-use Illuminate\Support\Facades\DB;
-use App\Models\DisbursementDetails;
-use App\Http\Controllers\Controller;
+use App\Models\WithdrawRequest;
+use App\Traits\Payment;
+use Carbon\Carbon;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
-use App\Library\Payment as PaymentInfo;
-use App\Mail\WithdrawRequestMail;
-use App\Models\Admin;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\Rules\Password;
-use App\Models\DisbursementWithdrawalMethod;
-use App\Models\WithdrawRequest;
-use Illuminate\Support\Facades\Mail;
 
 class DeliverymanController extends Controller
 {
-
     public function get_profile(Request $request)
     {
         $dm = DeliveryMan::with(['rating'])->where(['auth_token' => $request['token']])->first();
         $min_amount_to_pay_dm = BusinessSetting::where('key', 'min_amount_to_pay_dm')->first()->value ?? 0;
-        $dm['avg_rating'] = (float)(!empty($dm->rating[0]) ? $dm->rating[0]->average : 0);
-        $dm['rating_count'] = (float)(!empty($dm->rating[0]) ? $dm->rating[0]->rating_count : 0);
-        $dm['order_count'] = (int)$dm->orders->count();
-        $dm['todays_order_count'] = (int)$dm->todaysorders->count();
-        $dm['this_week_order_count'] = (int)$dm->this_week_orders->count();
-        $dm['member_since_days'] = (int)$dm->created_at->diffInDays();
+        $dm['avg_rating'] = (float) (! empty($dm->rating[0]) ? $dm->rating[0]->average : 0);
+        $dm['rating_count'] = (float) (! empty($dm->rating[0]) ? $dm->rating[0]->rating_count : 0);
+        $dm['order_count'] = (int) $dm->orders->count();
+        $dm['todays_order_count'] = (int) $dm->todaysorders->count();
+        $dm['this_week_order_count'] = (int) $dm->this_week_orders->count();
+        $dm['member_since_days'] = (int) $dm->created_at->diffInDays();
 
-        //Added DM TIPS
-        $dm['todays_earning'] = (float)($dm->todays_earning()->sum('original_delivery_charge') + $dm->todays_earning()->sum('dm_tips'));
-        $dm['this_week_earning'] = (float)($dm->this_week_earning()->sum('original_delivery_charge') + $dm->this_week_earning()->sum('dm_tips'));
-        $dm['this_month_earning'] = (float)($dm->this_month_earning()->sum('original_delivery_charge') + $dm->this_month_earning()->sum('dm_tips'));
+        // Added DM TIPS
+        $dm['todays_earning'] = (float) ($dm->todays_earning()->sum('original_delivery_charge') + $dm->todays_earning()->sum('dm_tips'));
+        $dm['this_week_earning'] = (float) ($dm->this_week_earning()->sum('original_delivery_charge') + $dm->this_week_earning()->sum('dm_tips'));
+        $dm['this_month_earning'] = (float) ($dm->this_month_earning()->sum('original_delivery_charge') + $dm->this_month_earning()->sum('dm_tips'));
 
         $dm['cash_in_hands'] = $dm->wallet ? $dm->wallet->collected_cash : 0;
         $dm['balance'] = $dm->wallet ? $dm->wallet->total_earning - ($dm->wallet->total_withdrawn + $dm?->wallet?->pending_withdraw) : 0;
         $dm['total_withdrawn'] = (float) ($dm?->wallet?->total_withdrawn ?? 0);
         $dm['total_earning'] = (float) ($dm?->wallet?->total_earning ?? 0);
         $dm['pending_withdraw'] = (float) ($dm?->wallet?->pending_withdraw ?? 0);
-        $dm['withdraw_able_balance'] = (float)($dm['balance'] - $dm?->wallet?->collected_cash > 0 ? abs($dm['balance'] - $dm?->wallet?->collected_cash) : 0);
-        $dm['Payable_Balance'] = (float)($dm?->wallet?->collected_cash ?? 0);
-
+        $dm['withdraw_able_balance'] = (float) ($dm['balance'] - $dm?->wallet?->collected_cash > 0 ? abs($dm['balance'] - $dm?->wallet?->collected_cash) : 0);
+        $dm['Payable_Balance'] = (float) ($dm?->wallet?->collected_cash ?? 0);
 
         $over_flow_balance = $dm['balance'] - $dm?->wallet?->collected_cash;
 
-        $wallet_earning =  round($dm?->wallet?->total_earning - ($dm?->wallet?->total_withdrawn +$dm?->wallet?->pending_withdraw) ,8);
-        if(isset($dm?->wallet) && (($over_flow_balance > 0 && $dm?->wallet?->collected_cash > 0 ) || ($dm?->wallet?->collected_cash != 0 && $dm['balance'] !=  0)) ){
+        $wallet_earning = round($dm?->wallet?->total_earning - ($dm?->wallet?->total_withdrawn + $dm?->wallet?->pending_withdraw), 8);
+        if (isset($dm?->wallet) && (($over_flow_balance > 0 && $dm?->wallet?->collected_cash > 0) || ($dm?->wallet?->collected_cash != 0 && $dm['balance'] != 0))) {
             $dm['adjust_able'] = true;
 
-        }  elseif( isset($dm?->wallet) &&  $over_flow_balance == $dm['balance']  ){
+        } elseif (isset($dm?->wallet) && $over_flow_balance == $dm['balance']) {
             $dm['adjust_able'] = false;
-        }
-        else{
+        } else {
             $dm['adjust_able'] = false;
         }
 
-        if($dm?->wallet?->collected_cash == 0 ||  $wallet_earning == 0 ){
+        if ($dm?->wallet?->collected_cash == 0 || $wallet_earning == 0) {
             $dm['adjust_able'] = false;
         }
 
@@ -92,20 +91,19 @@ class DeliverymanController extends Controller
             $dm['show_withdraw_button'] = true;
         }
 
-
-        $Payable_Balance =  $dm?->wallet?->collected_cash > 0 ? 1 : 0;
-        $cash_in_hand_overflow =  BusinessSetting::where('key', 'cash_in_hand_overflow_delivery_man')->first()?->value;
-        $cash_in_hand_overflow_delivery_man =  BusinessSetting::where('key', 'dm_max_cash_in_hand')->first()?->value;
-        $val =  $cash_in_hand_overflow_delivery_man - (($cash_in_hand_overflow_delivery_man * 10) / 100);
+        $Payable_Balance = $dm?->wallet?->collected_cash > 0 ? 1 : 0;
+        $cash_in_hand_overflow = BusinessSetting::where('key', 'cash_in_hand_overflow_delivery_man')->first()?->value;
+        $cash_in_hand_overflow_delivery_man = BusinessSetting::where('key', 'dm_max_cash_in_hand')->first()?->value;
+        $val = $cash_in_hand_overflow_delivery_man - (($cash_in_hand_overflow_delivery_man * 10) / 100);
         $dm['over_flow_warning'] = false;
-        $dm['dm_max_cash_in_hand'] =(float)$cash_in_hand_overflow_delivery_man ?? 0;
-        if ($Payable_Balance == 1 &&  $cash_in_hand_overflow &&  $over_flow_balance < 0 &&  $val <=  abs($dm?->wallet?->collected_cash)) {
+        $dm['dm_max_cash_in_hand'] = (float) $cash_in_hand_overflow_delivery_man ?? 0;
+        if ($Payable_Balance == 1 && $cash_in_hand_overflow && $over_flow_balance < 0 && $val <= abs($dm?->wallet?->collected_cash)) {
 
             $dm['over_flow_warning'] = true;
         }
 
         $dm['over_flow_block_warning'] = false;
-        if ($Payable_Balance == 1 &&  $cash_in_hand_overflow &&  $over_flow_balance < 0 &&  $cash_in_hand_overflow_delivery_man < abs($dm?->wallet?->collected_cash)) {
+        if ($Payable_Balance == 1 && $cash_in_hand_overflow && $over_flow_balance < 0 && $cash_in_hand_overflow_delivery_man < abs($dm?->wallet?->collected_cash)) {
             $dm['over_flow_block_warning'] = true;
         }
 
@@ -114,6 +112,7 @@ class DeliverymanController extends Controller
         unset($dm['todaysorders']);
         unset($dm['this_week_orders']);
         unset($dm['wallet']);
+
         return response()->json($dm, 200);
     }
 
@@ -123,7 +122,7 @@ class DeliverymanController extends Controller
         $validator = Validator::make($request->all(), [
             'f_name' => 'required',
             'l_name' => 'required',
-            'email' => 'required|unique:delivery_men,email,' . $dm->id,
+            'email' => 'required|unique:delivery_men,email,'.$dm->id,
             'password' => ['nullable', Password::min(8)->mixedCase()->letters()->numbers()->symbols()->uncompromised()],
         ], [
             'f_name.required' => 'First name is required!',
@@ -148,7 +147,7 @@ class DeliverymanController extends Controller
             $pass = $dm->password;
         }
 
-        $dm->vehicle_id = $request->vehicle_id ??  $dm->vehicle_id ?? null;
+        $dm->vehicle_id = $request->vehicle_id ?? $dm->vehicle_id ?? null;
 
         $dm->f_name = $request->f_name;
         $dm->l_name = $request->l_name;
@@ -175,6 +174,7 @@ class DeliverymanController extends Controller
         $dm = DeliveryMan::with(['rating'])->where(['auth_token' => $request['token']])->first();
         $dm->active = $dm->active ? 0 : 1;
         $dm->save();
+
         return response()->json(['message' => translate('messages.active_status_updated')], 200);
     }
 
@@ -189,6 +189,7 @@ class DeliverymanController extends Controller
             ->dmOrder()
             ->get();
         $orders = Helpers::order_data_formatting($orders, true);
+
         return response()->json($orders, 200);
     }
 
@@ -224,7 +225,7 @@ class DeliverymanController extends Controller
             $orders = $orders->where(function ($query) {
                 return $query->whereIn('order_status', ['confirmed', 'processing', 'handover'])
                     ->orWhere(function ($subQuery) {
-                        return  $subQuery->where('order_type', 'parcel')->whereIn('order_status', ['pending', 'confirmed', 'processing', 'handover']);
+                        return $subQuery->where('order_type', 'parcel')->whereIn('order_status', ['pending', 'confirmed', 'processing', 'handover']);
                     });
             });
         }
@@ -239,6 +240,7 @@ class DeliverymanController extends Controller
             ->orderBy('schedule_at', 'desc')
             ->get();
         $orders = Helpers::order_data_formatting($orders, true);
+
         return response()->json($orders, 200);
     }
 
@@ -256,43 +258,41 @@ class DeliverymanController extends Controller
             ->whereNull('delivery_man_id')
             ->dmOrder()
             ->first();
-        if (!$order) {
+        if (! $order) {
             return response()->json([
                 'errors' => [
-                    ['code' => 'order', 'message' => translate('messages.can_not_accept')]
-                ]
+                    ['code' => 'order', 'message' => translate('messages.can_not_accept')],
+                ],
             ], 404);
         }
         if ($dm->active != 1) {
             return response()->json([
                 'errors' => [
-                    ['code' => 'active_status', 'message' => translate('messages.You_can_not_accept_order_on_offline')]
-                ]
+                    ['code' => 'active_status', 'message' => translate('messages.You_can_not_accept_order_on_offline')],
+                ],
             ], 404);
         }
         if ($dm->current_orders >= config('dm_maximum_orders')) {
             return response()->json([
                 'errors' => [
-                    ['code' => 'dm_maximum_order_exceed', 'message' => translate('messages.dm_maximum_order_exceed_warning')]
-                ]
+                    ['code' => 'dm_maximum_order_exceed', 'message' => translate('messages.dm_maximum_order_exceed_warning')],
+                ],
             ], 405);
         }
 
         $payments = $order->payments()->where('payment_method', 'cash_on_delivery')->exists();
         $cash_in_hand = $dm?->wallet?->collected_cash ?? 0;
         $dm_max_cash = BusinessSetting::where('key', 'dm_max_cash_in_hand')->first();
-        $value =  $dm_max_cash?->value ?? 0;
+        $value = $dm_max_cash?->value ?? 0;
 
-
-        if (($order->payment_method == "cash_on_delivery" || $payments) && (($cash_in_hand + $order->order_amount) >= $value)) {
+        if (($order->payment_method == 'cash_on_delivery' || $payments) && (($cash_in_hand + $order->order_amount) >= $value)) {
 
             return response()->json([
                 'errors' => [
-                    ['code' => 'dm_maximum_hand_in_cash', 'message' => \App\CentralLogics\Helpers::format_currency($value) . " " . translate('max_cash_in_hand_exceeds')]
-                ]
+                    ['code' => 'dm_maximum_hand_in_cash', 'message' => \App\CentralLogics\Helpers::format_currency($value).' '.translate('max_cash_in_hand_exceeds')],
+                ],
             ], 405);
         }
-
 
         if ($order->order_type == 'parcel' && $order->order_status == 'confirmed') {
             $order->order_status = 'handover';
@@ -313,7 +313,6 @@ class DeliverymanController extends Controller
 
         $fcm_token = $order->is_guest == 0 ? $order?->customer?->cm_firebase_token : $order?->guest?->fcm_token;
 
-
         $value = Helpers::order_status_update_message('accepted', $order->module->module_type);
         $value = Helpers::text_variable_data_format(value: $value, store_name: $order->store?->name, order_id: $order->id, user_name: "{$order?->customer?->f_name} {$order?->customer?->l_name}", delivery_man_name: "{$order->delivery_man?->f_name} {$order->delivery_man?->l_name}");
         try {
@@ -323,7 +322,7 @@ class DeliverymanController extends Controller
                     'description' => $value,
                     'order_id' => $order['id'],
                     'image' => '',
-                    'type' => 'order_status'
+                    'type' => 'order_status',
                 ];
                 Helpers::send_push_notif_to_device($fcm_token, $data);
             }
@@ -344,14 +343,13 @@ class DeliverymanController extends Controller
             'created_at' => now(),
             'updated_at' => now()
         ]);
-
         return response()->json(['message' => translate('location recorded')], 200);
     }
 
     public function get_order_history(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'order_id' => 'required'
+            'order_id' => 'required',
         ]);
         if ($validator->fails()) {
             return response()->json(['errors' => Helpers::error_processor($validator)], 403);
@@ -359,13 +357,14 @@ class DeliverymanController extends Controller
         $dm = DeliveryMan::where(['auth_token' => $request['token']])->first();
 
         $history = DeliveryHistory::where(['order_id' => $request['order_id'], 'delivery_man_id' => $dm['id']])->get();
+
         return response()->json($history, 200);
     }
 
     public function send_order_otp(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'order_id' => 'required'
+            'order_id' => 'required',
         ]);
         if ($validator->fails()) {
             return response()->json(['errors' => Helpers::error_processor($validator)], 403);
@@ -380,14 +379,14 @@ class DeliverymanController extends Controller
             });
         }
         $order = $order->dmOrder()->first();
-        if (!$order) {
+        if (! $order) {
             return response()->json([
                 'errors' => [
-                    ['code' => 'order', 'message' => translate('messages.not_found')]
-                ]
+                    ['code' => 'order', 'message' => translate('messages.not_found')],
+                ],
             ], 404);
         }
-        $value = translate('your_order_is_ready_to_be_delivered,_plesae_share_your_otp_with_delivery_man.') . ' ' . translate('otp:') . $order->otp . ', ' . translate('order_id:') . $order->id;
+        $value = translate('your_order_is_ready_to_be_delivered,_plesae_share_your_otp_with_delivery_man.').' '.translate('otp:').$order->otp.', '.translate('order_id:').$order->id;
         try {
 
             $fcm_token = $order->is_guest == 0 ? $order?->customer?->cm_firebase_token : $order?->guest?->fcm_token;
@@ -405,13 +404,15 @@ class DeliverymanController extends Controller
                     'data' => json_encode($data),
                     'user_id' => $order->user_id,
                     'created_at' => now(),
-                    'updated_at' => now()
+                    'updated_at' => now(),
                 ]);
             }
         } catch (\Exception $e) {
             info($e->getMessage());
+
             return response()->json(['message' => translate('messages.push_notification_faild')], 403);
         }
+
         return response()->json([], 200);
     }
 
@@ -425,7 +426,7 @@ class DeliverymanController extends Controller
         ]);
 
         $validator->sometimes('otp', 'required', function ($request) {
-            return (Config::get('order_delivery_verification') == 1 && $request['status'] == 'delivered');
+            return Config::get('order_delivery_verification') == 1 && $request['status'] == 'delivered';
         });
 
         if ($validator->fails()) {
@@ -435,11 +436,11 @@ class DeliverymanController extends Controller
 
         $order = Order::where(['id' => $request['order_id'], 'delivery_man_id' => $dm['id']])->dmOrder()->first();
 
-        if (!$order || (!$order->store && $order->order_type != 'parcel')) {
+        if (! $order || (! $order->store && $order->order_type != 'parcel')) {
             return response()->json([
                 'errors' => [
-                    ['code' => 'not_found', 'message' => translate('messages.you_can_not_change_the_status_of_this_order')]
-                ]
+                    ['code' => 'not_found', 'message' => translate('messages.you_can_not_change_the_status_of_this_order')],
+                ],
             ], 403);
         }
 
@@ -448,45 +449,43 @@ class DeliverymanController extends Controller
             if (data_get($cancel_parcel_order, 'status_code') != 200) {
                 return response()->json([
                     'errors' => [
-                        ['code' => data_get($cancel_parcel_order, 'code'), 'message' => data_get($cancel_parcel_order, 'message')]
-                    ]
+                        ['code' => data_get($cancel_parcel_order, 'code'), 'message' => data_get($cancel_parcel_order, 'message')],
+                    ],
                 ], data_get($cancel_parcel_order, 'status_code'));
             } else {
                 return response()->json(['message' => translate('messages.Parcel_canceled_successfully')], 200);
             }
         }
 
-        if ($request['status'] == "confirmed" && config('order_confirmation_model') == 'store') {
+        if ($request['status'] == 'confirmed' && config('order_confirmation_model') == 'store') {
             return response()->json([
                 'errors' => [
-                    ['code' => 'order-confirmation-model', 'message' => translate('messages.order_confirmation_warning')]
-                ]
+                    ['code' => 'order-confirmation-model', 'message' => translate('messages.order_confirmation_warning')],
+                ],
             ], 403);
         }
 
-        if ($request['status'] == 'canceled' && !config('canceled_by_deliveryman')) {
+        if ($request['status'] == 'canceled' && ! config('canceled_by_deliveryman')) {
             return response()->json([
                 'errors' => [
-                    ['code' => 'status', 'message' => translate('messages.you_can_not_cancel_a_order')]
-                ]
+                    ['code' => 'status', 'message' => translate('messages.you_can_not_cancel_a_order')],
+                ],
             ], 403);
         }
 
         if ($order->confirmed && $request['status'] == 'canceled') {
             return response()->json([
                 'errors' => [
-                    ['code' => 'delivery-man', 'message' => translate('messages.order_can_not_cancle_after_confirm')]
-                ]
+                    ['code' => 'delivery-man', 'message' => translate('messages.order_can_not_cancle_after_confirm')],
+                ],
             ], 403);
         }
 
-
-
-        if (Config::get('order_delivery_verification') == 1 &&  $request['status'] == 'delivered' && $order->otp != $request['otp']) {
+        if (Config::get('order_delivery_verification') == 1 && $request['status'] == 'delivered' && $order->otp != $request['otp']) {
             return response()->json([
                 'errors' => [
-                    ['code' => 'otp', 'message' => translate('Otp Not matched')]
-                ]
+                    ['code' => 'otp', 'message' => translate('Otp Not matched')],
+                ],
             ], 406);
         }
         if ($request->status == 'delivered') {
@@ -503,8 +502,8 @@ class DeliverymanController extends Controller
                 } else {
                     return response()->json([
                         'errors' => [
-                            ['code' => 'error', 'message' => translate('messages.faield_to_create_order_transaction')]
-                        ]
+                            ['code' => 'error', 'message' => translate('messages.faield_to_create_order_transaction')],
+                        ],
                     ], 406);
                 }
             }
@@ -532,7 +531,7 @@ class DeliverymanController extends Controller
 
             $img_names = [];
             $images = [];
-            if (!empty($request->file('order_proof'))) {
+            if (! empty($request->file('order_proof'))) {
                 foreach ($request->order_proof as $img) {
                     $image_name = Helpers::upload('order/', 'png', $img);
                     array_push($img_names, ['img' => $image_name, 'storage' => Helpers::getDisk()]);
@@ -544,7 +543,7 @@ class DeliverymanController extends Controller
             }
 
             OrderLogic::update_unpaid_order_payment(order_id: $order->id, payment_method: $order->payment_method);
-        } else if ($request->status == 'canceled') {
+        } elseif ($request->status == 'canceled') {
             if ($order->delivery_man) {
                 $dm = $order->delivery_man;
                 $dm->current_orders = $dm->current_orders > 1 ? $dm->current_orders - 1 : 0;
@@ -552,11 +551,10 @@ class DeliverymanController extends Controller
             }
             $order->cancellation_reason = $request->reason;
             $order->canceled_by = 'deliveryman';
-        } else if ($order->order_type == 'parcel' && $request->status == 'handover') {
+        } elseif ($order->order_type == 'parcel' && $request->status == 'handover') {
             $order->confirmed = now();
             $order->processing = now();
-        }
-         else if ($order->order_type != 'parcel' && in_array($request->status, ['picked_up']) ) {
+        } elseif ($order->order_type != 'parcel' && in_array($request->status, ['picked_up'])) {
             Helpers::sendOrderDeliveryVerificationOtp($order);
         }
 
@@ -566,13 +564,13 @@ class DeliverymanController extends Controller
 
         Helpers::send_order_notification($order);
 
-        return response()->json(['message' =>  translate('Status updated')], 200);
+        return response()->json(['message' => translate('Status updated')], 200);
     }
 
     public function get_order_details(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'order_id' => 'required'
+            'order_id' => 'required',
         ]);
         if ($validator->fails()) {
             return response()->json(['errors' => Helpers::error_processor($validator)], 403);
@@ -584,21 +582,23 @@ class DeliverymanController extends Controller
             $query->WhereNull('delivery_man_id')
                 ->orWhere('delivery_man_id', $dm['id']);
         })->Notpos()->first();
-        if (!$order) {
+        if (! $order) {
             return response()->json([
                 'errors' => [
-                    ['code' => 'order', 'message' => translate('messages.not_found')]
-                ]
+                    ['code' => 'order', 'message' => translate('messages.not_found')],
+                ],
             ], 404);
         }
         $details = isset($order->details) ? $order->details : null;
         if ($details != null && $details->count() > 0) {
             $details[0]['vendor_id'] = $order?->store?->vendor_id;
             $details = $details = Helpers::order_details_data_formatting($details);
-            $details[0]['is_guest'] = (int)$order->is_guest;
+            $details[0]['is_guest'] = (int) $order->is_guest;
+
             return response()->json($details, 200);
-        } else if ($order->order_type == 'parcel') {
+        } elseif ($order->order_type == 'parcel') {
             $order->delivery_address = $order->delivery_address ? json_decode($order->delivery_address, true) : [];
+
             return response()->json(($order), 200);
         } elseif ($order->prescription_order == 1) {
             return response()->json([], 200);
@@ -606,29 +606,30 @@ class DeliverymanController extends Controller
 
         return response()->json([
             'errors' => [
-                ['code' => 'order', 'message' => translate('messages.not_found')]
-            ]
+                ['code' => 'order', 'message' => translate('messages.not_found')],
+            ],
         ], 404);
     }
 
     public function get_order(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'order_id' => 'required'
+            'order_id' => 'required',
         ]);
         if ($validator->fails()) {
             return response()->json(['errors' => Helpers::error_processor($validator)], 403);
         }
         $dm = DeliveryMan::where(['auth_token' => $request['token']])->first();
 
-        $order = Order::with(['customer', 'store', 'details', 'parcel_category', 'payments','ParcelCancellation'])->where(['delivery_man_id' => $dm['id'], 'id' => $request['order_id']])->Notpos()->first();
-        if (!$order) {
+        $order = Order::with(['customer', 'store', 'details', 'parcel_category', 'payments', 'ParcelCancellation'])->where(['delivery_man_id' => $dm['id'], 'id' => $request['order_id']])->Notpos()->first();
+        if (! $order) {
             return response()->json([
                 'errors' => [
-                    ['code' => 'order', 'message' => translate('messages.not_found')]
-                ]
+                    ['code' => 'order', 'message' => translate('messages.not_found')],
+                ],
             ], 204);
         }
+
         return response()->json(Helpers::order_data_formatting($order), 200);
     }
 
@@ -656,15 +657,16 @@ class DeliverymanController extends Controller
             'total_size' => $paginator->total(),
             'limit' => $request['limit'],
             'offset' => $request['offset'],
-            'orders' => $orders
+            'orders' => $orders,
         ];
+
         return response()->json($data, 200);
     }
 
     public function get_last_location(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'order_id' => 'required'
+            'order_id' => 'required',
         ]);
         if ($validator->fails()) {
             return response()->json(['errors' => Helpers::error_processor($validator)], 403);
@@ -673,6 +675,7 @@ class DeliverymanController extends Controller
         $last_data = DeliveryHistory::whereHas('delivery_man.orders', function ($query) use ($request) {
             return $query->where('id', $request->order_id);
         })->latest()->first();
+
         return response()->json($last_data, 200);
     }
 
@@ -680,7 +683,7 @@ class DeliverymanController extends Controller
     {
         $validator = Validator::make($request->all(), [
             'order_id' => 'required',
-            'status' => 'required|in:paid'
+            'status' => 'required|in:paid',
         ]);
         if ($validator->fails()) {
             return response()->json(['errors' => Helpers::error_processor($validator)], 403);
@@ -690,21 +693,23 @@ class DeliverymanController extends Controller
 
         if (Order::where(['delivery_man_id' => $dm['id'], 'id' => $request['order_id']])->dmOrder()->first()) {
             Order::where(['delivery_man_id' => $dm['id'], 'id' => $request['order_id']])->update([
-                'payment_status' => $request['status']
+                'payment_status' => $request['status'],
             ]);
+
             return response()->json(['message' => translate('Payment status updated')], 200);
         }
+
         return response()->json([
             'errors' => [
-                ['code' => 'order', 'message' => translate('not found!')]
-            ]
+                ['code' => 'order', 'message' => translate('not found!')],
+            ],
         ], 404);
     }
 
     public function update_fcm_token(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'fcm_token' => 'required'
+            'fcm_token' => 'required',
         ]);
         if ($validator->fails()) {
             return response()->json(['errors' => Helpers::error_processor($validator)], 403);
@@ -712,7 +717,7 @@ class DeliverymanController extends Controller
         $dm = DeliveryMan::where(['auth_token' => $request['token']])->first();
 
         DeliveryMan::where(['id' => $dm['id']])->update([
-            'fcm_token' => $request['fcm_token']
+            'fcm_token' => $request['fcm_token'],
         ]);
 
         return response()->json(['message' => translate('successfully updated!')], 200);
@@ -731,7 +736,7 @@ class DeliverymanController extends Controller
 
         $notifications->append('data');
 
-        $notifications =  $notifications->merge($user_notifications);
+        $notifications = $notifications->merge($user_notifications);
         try {
             return response()->json($notifications, 200);
         } catch (\Exception $e) {
@@ -751,9 +756,7 @@ class DeliverymanController extends Controller
             return response()->json(['errors' => [['code' => 'on-going', 'message' => translate('messages.You_have_cash_in_hand,_you_have_to_pay_the_due_to_delete_your_account.')]]], 203);
         }
 
-
         Helpers::check_and_delete('delivery-man/', $dm['image']);
-
 
         foreach (json_decode($dm['identity_image'], true) as $img) {
             Helpers::check_and_delete('delivery-man/', $img);
@@ -763,17 +766,18 @@ class DeliverymanController extends Controller
             $dm->userinfo->delete();
         }
         $dm->delete();
+
         return response()->json([]);
     }
+
     public function make_payment(Request $request)
     {
         $validator = Validator::make($request->all(), [
             'payment_gateway' => 'required',
             'amount' => 'required|numeric|min:.001',
             'callback' => 'required',
-            'token' => 'required'
+            'token' => 'required',
         ]);
-
 
         if ($validator->fails()) {
             return response()->json(['errors' => Helpers::error_processor($validator)], 403);
@@ -790,7 +794,7 @@ class DeliverymanController extends Controller
         $store_logo = BusinessSetting::where(['key' => 'logo'])->first();
         $additional_data = [
             'business_name' => BusinessSetting::where(['key' => 'business_name'])->first()?->value,
-            'business_logo' => \App\CentralLogics\Helpers::get_full_url('business', $store_logo?->value, $store_logo?->storage[0]?->value ?? 'public')
+            'business_logo' => \App\CentralLogics\Helpers::get_full_url('business', $store_logo?->value, $store_logo?->storage[0]?->value ?? 'public'),
         ];
         $payment_info = new PaymentInfo(
             success_hook: 'collect_cash_success',
@@ -813,9 +817,9 @@ class DeliverymanController extends Controller
         $data = [
             'redirect_link' => $redirect_link,
         ];
+
         return response()->json($data, 200);
     }
-
 
     public function make_wallet_adjustment(Request $request)
     {
@@ -824,40 +828,40 @@ class DeliverymanController extends Controller
         $wallet = DeliveryManWallet::firstOrNew(
             ['delivery_man_id' => $dm->id]
         );
-        $wallet_earning =  round($wallet->total_earning - ($wallet->total_withdrawn + $wallet->pending_withdraw), 8);
-        $adj_amount =  $wallet->collected_cash - $wallet_earning;
+        $wallet_earning = round($wallet->total_earning - ($wallet->total_withdrawn + $wallet->pending_withdraw), 8);
+        $adj_amount = $wallet->collected_cash - $wallet_earning;
 
         if ($wallet->collected_cash == 0 || $wallet_earning == 0) {
             return response()->json(['message' => translate('messages.Already_Adjusted')], 201);
         }
 
         if ($adj_amount > 0) {
-            $wallet->total_withdrawn =  $wallet->total_withdrawn + $wallet_earning;
-            $wallet->collected_cash =   $wallet->collected_cash - $wallet_earning;
+            $wallet->total_withdrawn = $wallet->total_withdrawn + $wallet_earning;
+            $wallet->collected_cash = $wallet->collected_cash - $wallet_earning;
 
             $data = [
                 'delivery_man_id' => $dm->id,
                 'amount' => $wallet_earning,
-                'ref' => "delivery_man_wallet_adjustment_partial",
-                'method' => "adjustment",
+                'ref' => 'delivery_man_wallet_adjustment_partial',
+                'method' => 'adjustment',
                 // 'approved' => 1,
                 // 'type' => 'adjustment',
                 'created_at' => now(),
-                'updated_at' => now()
+                'updated_at' => now(),
             ];
         } else {
             $data = [
                 'delivery_man_id' => $dm->id,
                 'amount' => $wallet->collected_cash,
-                'ref' => "delivery_man_wallet_adjustment_full",
-                'method' => "adjustment",
+                'ref' => 'delivery_man_wallet_adjustment_full',
+                'method' => 'adjustment',
                 // 'approved' => 1,
                 // 'type' => 'adjustment',
                 'created_at' => now(),
-                'updated_at' => now()
+                'updated_at' => now(),
             ];
-            $wallet->total_withdrawn =  $wallet->total_withdrawn + $wallet->collected_cash;
-            $wallet->collected_cash =   0;
+            $wallet->total_withdrawn = $wallet->total_withdrawn + $wallet->collected_cash;
+            $wallet->collected_cash = 0;
         }
 
         $wallet->save();
@@ -865,6 +869,7 @@ class DeliverymanController extends Controller
 
         return response()->json(['message' => translate('messages.Delivery_man_wallet_adjustment_successfull')], 200);
     }
+
     public function wallet_payment_list(Request $request)
     {
         $limit = $request['limit'] ?? 25;
@@ -873,12 +878,12 @@ class DeliverymanController extends Controller
 
         $key = isset($request['search']) ? explode(' ', $request['search']) : [];
         $paginator = AccountTransaction::when(isset($key), function ($query) use ($key) {
-                return $query->where(function ($q) use ($key) {
-                    foreach ($key as $value) {
-                        $q->orWhere('ref', 'like', "%{$value}%");
-                    }
-                });
-            })
+            return $query->where(function ($q) use ($key) {
+                foreach ($key as $value) {
+                    $q->orWhere('ref', 'like', "%{$value}%");
+                }
+            });
+        })
             ->where('type', 'collected')
             ->where('created_by', 'deliveryman')
             ->where('from_id', $dm->id)
@@ -904,6 +909,7 @@ class DeliverymanController extends Controller
 
         return response()->json($data, 200);
     }
+
     public function wallet_provided_earning_list(Request $request)
     {
         $limit = $request['limit'] ?? 25;
@@ -912,12 +918,12 @@ class DeliverymanController extends Controller
 
         $key = isset($request['search']) ? explode(' ', $request['search']) : [];
         $paginator = ProvideDMEarning::when(isset($key), function ($query) use ($key) {
-                return $query->where(function ($q) use ($key) {
-                    foreach ($key as $value) {
-                        $q->orWhere('ref', 'like', "%{$value}%");
-                    }
-                });
-            })
+            return $query->where(function ($q) use ($key) {
+                foreach ($key as $value) {
+                    $q->orWhere('ref', 'like', "%{$value}%");
+                }
+            });
+        })
             ->where('delivery_man_id', $dm->id)
             ->where('method', 'adjustment')
             ->whereIn('ref', ['delivery_man_wallet_adjustment_partial', 'delivery_man_wallet_adjustment_full'])
@@ -992,14 +998,16 @@ class DeliverymanController extends Controller
             'total_size' => $paginator->total(),
             'limit' => $request['limit'],
             'offset' => $request['offset'],
-            'methods' => $datas
+            'methods' => $datas,
         ];
+
         return response()->json($data, 200);
     }
 
     public function withdraw_method_list()
     {
         $wi = WithdrawalMethod::where('is_active', 1)->get();
+
         return response()->json($wi, 200);
     }
 
@@ -1021,20 +1029,20 @@ class DeliverymanController extends Controller
 
         $method_data = [];
         foreach ($fields as $field) {
-            if (key_exists($field, $values)) {
+            if (array_key_exists($field, $values)) {
                 $method_data[$field] = $values[$field];
             }
         }
 
-            $disbursementMethod = DisbursementWithdrawalMethod::firstOrNew([
-                'id' => $request->disbursement_withdrawal_method_id ?? null
-            ]);
-            $disbursementMethod->delivery_man_id      = $dm['id'];
-            $disbursementMethod->withdrawal_method_id = $method['id'];
-            $disbursementMethod->method_name          = $method['method_name'];
-            $disbursementMethod->method_fields        = json_encode($method_data);
-            $disbursementMethod->is_default = $disbursementMethod->exists  ? $disbursementMethod->is_default : 0;
-            $disbursementMethod->save();
+        $disbursementMethod = DisbursementWithdrawalMethod::firstOrNew([
+            'id' => $request->disbursement_withdrawal_method_id ?? null,
+        ]);
+        $disbursementMethod->delivery_man_id = $dm['id'];
+        $disbursementMethod->withdrawal_method_id = $method['id'];
+        $disbursementMethod->method_name = $method['method_name'];
+        $disbursementMethod->method_fields = json_encode($method_data);
+        $disbursementMethod->is_default = $disbursementMethod->exists ? $disbursementMethod->is_default : 0;
+        $disbursementMethod->save();
 
         return response()->json(['message' => translate('successfully added!')], 200);
     }
@@ -1054,6 +1062,7 @@ class DeliverymanController extends Controller
         $method->is_default = $request->is_default;
         $method->save();
         DisbursementWithdrawalMethod::whereNot('id', $request->id)->where('delivery_man_id', $dm['id'])->update(['is_default' => 0]);
+
         return response()->json(['message' => translate('messages.method_updated_successfully')], 200);
     }
 
@@ -1069,6 +1078,7 @@ class DeliverymanController extends Controller
 
         $method = DisbursementWithdrawalMethod::find($request->id);
         $method->delete();
+
         return response()->json(['message' => translate('messages.method_deleted_successfully')], 200);
     }
 
@@ -1091,7 +1101,7 @@ class DeliverymanController extends Controller
         $paginator = DisbursementDetails::where('delivery_man_id', $dm['id'])->latest()->paginate($limit, ['*'], 'page', $offset);
 
         $paginator->each(function ($data) {
-            $data->withdraw_method?->method_fields ?  $data->withdraw_method->method_fields = json_decode($data->withdraw_method?->method_fields, true) : '';
+            $data->withdraw_method?->method_fields ? $data->withdraw_method->method_fields = json_decode($data->withdraw_method?->method_fields, true) : '';
         });
 
         $data = [
@@ -1102,10 +1112,12 @@ class DeliverymanController extends Controller
             'completed' => (float) $total_disbursements->where('status', 'completed')->sum('disbursement_amount'),
             'canceled' => (float) $total_disbursements->where('status', 'canceled')->sum('disbursement_amount'),
             'complete_day' => (int) BusinessSetting::where(['key' => 'dm_disbursement_waiting_time'])->first()?->value,
-            'disbursements' => $paginator->items()
+            'disbursements' => $paginator->items(),
         ];
+
         return response()->json($data, 200);
     }
+
     public function earningReport(Request $request)
     {
         $validator = Validator::make($request->all(), [
@@ -1164,15 +1176,15 @@ class DeliverymanController extends Controller
             $item->dm_tips = (float) $item->dm_tips;
             $item->original_delivery_charge = (float) $item->original_delivery_charge;
             $item->delivery_fee_comission = (float) $item->delivery_fee_comission;
+
             return $item;
         });
 
-
         $data = [
             'earning' => $paginated_orders,
-            'total_dm_tips' => (float)$total_dm_tips,
-            'total_delivery_charge' => (float)$total_delivery_charge,
-            'total_admin_commission' => (float)$total_admin_commission,
+            'total_dm_tips' => (float) $total_dm_tips,
+            'total_delivery_charge' => (float) $total_delivery_charge,
+            'total_admin_commission' => (float) $total_admin_commission,
             'type' => $type,
             'limit' => $limit,
             'offset' => $offset,
@@ -1191,7 +1203,7 @@ class DeliverymanController extends Controller
         $status = [
             0 => 'Pending',
             1 => 'Approved',
-            2 => 'Denied'
+            2 => 'Denied',
         ];
         foreach ($withdraw_req as $item) {
             $item['status'] = $status[$item->approved];
@@ -1217,7 +1229,7 @@ class DeliverymanController extends Controller
     {
         $validator = Validator::make($request->all(), [
             'amount' => 'required|numeric|min:0.01',
-            'id' => 'required'
+            'id' => 'required',
         ]);
         if ($validator->fails()) {
             return response()->json(['errors' => Helpers::error_processor($validator)], 403);
@@ -1231,7 +1243,7 @@ class DeliverymanController extends Controller
 
         $method_data = [];
         foreach ($fields as $field) {
-            if (key_exists($field, $values)) {
+            if (array_key_exists($field, $values)) {
                 $method_data[$field] = $values[$field];
             }
         }
@@ -1247,31 +1259,32 @@ class DeliverymanController extends Controller
                 'withdrawal_method_fields' => json_encode($method_data),
                 'approved' => 0,
                 'created_at' => now(),
-                'updated_at' => now()
+                'updated_at' => now(),
             ];
             try {
                 DB::table('withdraw_requests')->insert($data);
                 $w?->increment('pending_withdraw', $request['amount']);
                 $mail_status = Helpers::get_mail_status('dm_withdraw_request_mail_status_admin');
-                $admin= Admin::where('role_id', 1)->first();
-                $wallet_transaction = WithdrawRequest::where('delivery_man_id',$w->delivery_man_id)->latest()->first();
-                if(  config('mail.status') && $mail_status == '1'  &&  Helpers::getNotificationStatusData('admin','dm_withdraw_request','mail_status' )) {
-                    Mail::to($admin->email)->send(new WithdrawRequestMail('admin_mail',$wallet_transaction,'dm'));
+                $admin = Admin::where('role_id', 1)->first();
+                $wallet_transaction = WithdrawRequest::where('delivery_man_id', $w->delivery_man_id)->latest()->first();
+                if (config('mail.status') && $mail_status == '1' && Helpers::getNotificationStatusData('admin', 'dm_withdraw_request', 'mail_status')) {
+                    Mail::to($admin->email)->send(new WithdrawRequestMail('admin_mail', $wallet_transaction, 'dm'));
                 }
 
                 return response()->json(['message' => translate('messages.withdraw_request_placed_successfully')], 200);
             } catch (\Exception $e) {
                 info($e->getMessage());
+
                 return response()->json($e);
             }
         }
+
         return response()->json([
             'errors' => [
-                ['code' => 'amount', 'message' => translate('messages.insufficient_balance')]
-            ]
+                ['code' => 'amount', 'message' => translate('messages.insufficient_balance')],
+            ],
         ], 403);
     }
-
 
     public function addReturnDate(Request $request)
     {
@@ -1289,19 +1302,20 @@ class DeliverymanController extends Controller
             $order->ParcelCancellation->return_date = $request['return_date'];
             $order->ParcelCancellation->set_return_date = 1;
             $order->ParcelCancellation->save();
+
             return response()->json([
-                'message' => translate('messages.return_date_added_successfully')
+                'message' => translate('messages.return_date_added_successfully'),
             ], 200);
         } else {
             return response()->json([
                 'errors' => [
-                    ['code' => 'amount', 'message' => translate('Order not found')]
-                ]
+                    ['code' => 'amount', 'message' => translate('Order not found')],
+                ],
             ], 403);
         }
     }
 
-        public function parcelReturn(Request $request)
+    public function parcelReturn(Request $request)
     {
         $validator = Validator::make($request->all(), [
             'order_id' => 'required',
@@ -1315,23 +1329,22 @@ class DeliverymanController extends Controller
         $dmId = DeliveryMan::where(['auth_token' => $request['token']])->first()?->id;
         $order = Order::where('id', $request['order_id'])->where('delivery_man_id', $dmId)->with('ParcelCancellation')->first();
 
-        $validationCheck =  OrderLogic::makeValidationForParcelReturn($request,$order);
+        $validationCheck = OrderLogic::makeValidationForParcelReturn($request, $order);
         if (data_get($validationCheck, 'status_code') === 403) {
 
             return response()->json([
                 'errors' => [
-                    ['code' => data_get($validationCheck, 'code'), 'message' => data_get($validationCheck, 'message')]
-                ]
+                    ['code' => data_get($validationCheck, 'code'), 'message' => data_get($validationCheck, 'message')],
+                ],
             ], data_get($validationCheck, 'status_code'));
         }
 
-        if( in_array($order->parcelCancellation->cancel_by ,['deliveryman', 'admin_for_deliveryman']  )){
-            OrderLogic::deliveryManCancelParcelTransaction($order,'deliveryman');
-        } else{
-            OrderLogic::create_transaction_parcel_cancel($order, $order->payment_status == 'paid' ? 'admin' : 'deliveryman', );
+        if (in_array($order->parcelCancellation->cancel_by, ['deliveryman', 'admin_for_deliveryman'])) {
+            OrderLogic::deliveryManCancelParcelTransaction($order, 'deliveryman');
+        } else {
+            OrderLogic::create_transaction_parcel_cancel($order, $order->payment_status == 'paid' ? 'admin' : 'deliveryman');
         }
 
         return response()->json(['message' => translate('messages.Parcel_returned_successfully')], 200);
     }
-
 }
