@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\ReactPromotionalBanner;
 use App\Models\Store;
 use App\Models\Setting;
 use App\Models\DataSetting;
@@ -42,15 +43,10 @@ class UpdateController extends Controller
         Helpers::setEnvironmentValue('BUYER_USERNAME', $request['username']);
         Helpers::setEnvironmentValue('PURCHASE_CODE', $request['purchase_key']);
         Helpers::setEnvironmentValue('APP_MODE', 'live');
-        Helpers::setEnvironmentValue('SOFTWARE_VERSION', '3.4');
+        Helpers::setEnvironmentValue('SOFTWARE_VERSION', '3.5');
         Helpers::setEnvironmentValue('REACT_APP_KEY', '45370351');
         Helpers::setEnvironmentValue('APP_NAME', '6amMart' . time());
 
-        // $data = Helpers::requestSender();
-        // if (!$data['active']) {
-        // if (!$this->actch()) {
-        //     return redirect(base64_decode('aHR0cHM6Ly82YW10ZWNoLmNvbS9zb2Z0d2FyZS1hY3RpdmF0aW9u'));
-        // }
 
         // version_2.11.1
         Artisan::call('cache:table');
@@ -62,9 +58,9 @@ class UpdateController extends Controller
         copy($newRouteServiceProvier, $previousRouteServiceProvier);
         Artisan::call('cache:clear');
         Artisan::call('view:clear');
-        Helpers::insert_business_settings_key("mobile_app_section_heading" , "Download the App for Enjoy Best Restaurant Test");
-        Helpers::insert_business_settings_key("mobile_app_section_text" , "Default Text Mobile App Section");
-        Helpers::insert_business_settings_key("feature_section_description" , "Feature section description");
+        Helpers::insert_business_settings_key("mobile_app_section_heading", "Download the App for Enjoy Best Restaurant Test");
+        Helpers::insert_business_settings_key("mobile_app_section_text", "Default Text Mobile App Section");
+        Helpers::insert_business_settings_key("feature_section_description", "Feature section description");
         Helpers::insert_business_settings_key("Feature section description", json_encode([
             "app_url_android_status" => "0",
             "app_url_android" => "https://play.google.com",
@@ -75,42 +71,42 @@ class UpdateController extends Controller
         ]));
 
         //version 1.5.0
-        Helpers::insert_business_settings_key("wallet_status" , "0");
-        Helpers::insert_business_settings_key("loyalty_point_status" , "0");
-        Helpers::insert_business_settings_key("ref_earning_status" , "0");
-        Helpers::insert_business_settings_key("wallet_add_refund" , "0");
-        Helpers::insert_business_settings_key("loyalty_point_exchange_rate" , "0");
-        Helpers::insert_business_settings_key("ref_earning_exchange_rate" , "0");
-        Helpers::insert_business_settings_key("loyalty_point_item_purchase_point" , "0");
-        Helpers::insert_business_settings_key("loyalty_point_minimum_point" , "0");
-        Helpers::insert_business_settings_key("dm_tips_status" , "0");
+        Helpers::insert_business_settings_key("wallet_status", "0");
+        Helpers::insert_business_settings_key("loyalty_point_status", "0");
+        Helpers::insert_business_settings_key("ref_earning_status", "0");
+        Helpers::insert_business_settings_key("wallet_add_refund", "0");
+        Helpers::insert_business_settings_key("loyalty_point_exchange_rate", "0");
+        Helpers::insert_business_settings_key("ref_earning_exchange_rate", "0");
+        Helpers::insert_business_settings_key("loyalty_point_item_purchase_point", "0");
+        Helpers::insert_business_settings_key("loyalty_point_minimum_point", "0");
+        Helpers::insert_business_settings_key("dm_tips_status", "0");
 
         Helpers::insert_business_settings_key('refund_active_status', '1');
-        Helpers::insert_business_settings_key('social_login','[{"login_medium":"google","client_id":"","client_secret":"","status":"0"},{"login_medium":"facebook","client_id":"","client_secret":"","status":""}]');
-        Helpers::insert_business_settings_key('system_language','[{"id":1,"direction":"ltr","code":"en","status":1,"default":true}]');
-        Helpers::insert_business_settings_key('language','["en"]');
+        Helpers::insert_business_settings_key('social_login', '[{"login_medium":"google","client_id":"","client_secret":"","status":"0"},{"login_medium":"facebook","client_id":"","client_secret":"","status":""}]');
+        Helpers::insert_business_settings_key('system_language', '[{"id":1,"direction":"ltr","code":"en","status":1,"default":true}]');
+        Helpers::insert_business_settings_key('language', '["en"]');
         //version 2.0.1
         // Helpers::insert_business_settings_key('otp_interval_time', '30');
         // Helpers::insert_business_settings_key('max_otp_hit', '5');
 
-        Helpers::insert_business_settings_key("home_delivery_status" , "1");
-        Helpers::insert_business_settings_key("takeaway_status" , "1");
+        Helpers::insert_business_settings_key("home_delivery_status", "1");
+        Helpers::insert_business_settings_key("takeaway_status", "1");
 
         $data_settings = file_get_contents('database/partial/data_settings.sql');
         $email_tempaltes = file_get_contents('database/partial/email_tempaltes.sql');
 
-        if( DataSetting::count() < 1){
+        if (DataSetting::count() < 1) {
             DB::statement($data_settings);
         }
-        if( EmailTemplate::count() < 1){
+        if (EmailTemplate::count() < 1) {
             DB::statement($email_tempaltes);
         }
 
         //version 2.2.0
-        Helpers::insert_data_settings_key('admin_login_url', 'login_admin' ,'admin');
-        Helpers::insert_data_settings_key('admin_employee_login_url', 'login_admin_employee' ,'admin-employee');
-        Helpers::insert_data_settings_key('store_login_url', 'login_store' ,'vendor');
-        Helpers::insert_data_settings_key('store_employee_login_url', 'login_store_employee' ,'vendor-employee');
+        Helpers::insert_data_settings_key('admin_login_url', 'login_admin', 'admin');
+        Helpers::insert_data_settings_key('admin_employee_login_url', 'login_admin_employee', 'admin-employee');
+        Helpers::insert_data_settings_key('store_login_url', 'login_store', 'vendor');
+        Helpers::insert_data_settings_key('store_employee_login_url', 'login_store_employee', 'vendor-employee');
 
         Helpers::insert_business_settings_key('subscription_business_model', '0');
         Helpers::insert_business_settings_key('commission_business_model', '1');
@@ -144,10 +140,10 @@ class UpdateController extends Controller
             }
 
 
-            $storesToUpdate = Store::whereNull('slug')->get(['id','name','slug']);
+            $storesToUpdate = Store::whereNull('slug')->get(['id', 'name', 'slug']);
             foreach ($storesToUpdate as $store) {
                 $slug = Str::slug($store->name);
-                $store->slug = $store->slug? $store->slug :"{$slug}{$store->id}";
+                $store->slug = $store->slug ? $store->slug : "{$slug}{$store->id}";
                 $store->save();
             }
 
@@ -160,7 +156,7 @@ class UpdateController extends Controller
                 if ($data_values) {
                     $additional_data = $data_values->live_values;
 
-                    if (array_key_exists("callback_url",$additional_data)) {
+                    if (array_key_exists("callback_url", $additional_data)) {
                         unset($additional_data['callback_url']);
                         $data_values->live_values = $additional_data;
                         $data_values->test_values = $additional_data;
@@ -176,9 +172,9 @@ class UpdateController extends Controller
                 ->limit(1)
                 ->exists();
 
-                if($hasDuplicates){
-                    DeliveryHistory::truncate();
-                }
+            if ($hasDuplicates) {
+                DeliveryHistory::truncate();
+            }
 
         } catch (\Exception $exception) {
             Toastr::error('Database import failed! try again');
@@ -186,13 +182,13 @@ class UpdateController extends Controller
         }
 
         $landing = BusinessSetting::where('key', 'landing_page')->exists();
-        if(!$landing){
-            Helpers::insert_business_settings_key('landing_page','1');
-            Helpers::insert_business_settings_key('landing_integration_type','none');
+        if (!$landing) {
+            Helpers::insert_business_settings_key('landing_page', '1');
+            Helpers::insert_business_settings_key('landing_integration_type', 'none');
         }
-        Helpers::insert_business_settings_key("dm_max_cash_in_hand" , "5000");
+        Helpers::insert_business_settings_key("dm_max_cash_in_hand", "5000");
 
-        if(NotificationSetting::count() == 0 ){
+        if (NotificationSetting::count() == 0) {
             Helpers::notificationDataSetup();
         }
         Helpers::updateAdminNotificationSetupDataSetup();
@@ -203,9 +199,9 @@ class UpdateController extends Controller
 
         $this->firebase_message_config_file_gen();
 
-        $recaptcha= BusinessSetting::where('key','recaptcha')->first();
-        if($recaptcha?->value){
-            $recaptcha_value=  json_decode($recaptcha->value,true);
+        $recaptcha = BusinessSetting::where('key', 'recaptcha')->first();
+        if ($recaptcha?->value) {
+            $recaptcha_value = json_decode($recaptcha->value, true);
             $recaptcha->value = json_encode([
                 'status' => null,
                 'site_key' => $recaptcha_value['site_key'],
@@ -213,12 +209,13 @@ class UpdateController extends Controller
             ]);
             $recaptcha->save();
         }
+        Helpers::promotionalImage();
 
         Coupon::where('coupon_type', 'store_wise')
-        ->whereNull('store_id')
-        ->update([
-            'store_id' => DB::raw("JSON_UNQUOTE(JSON_EXTRACT(data, '$[0]'))")
-        ]);
+            ->whereNull('store_id')
+            ->update([
+                'store_id' => DB::raw("JSON_UNQUOTE(JSON_EXTRACT(data, '$[0]'))")
+            ]);
 
         $twoFactor = Setting::where(['key_name' => '2factor', 'settings_type' => 'sms_config'])->first();
         if ($twoFactor && $twoFactor->live_values) {
@@ -230,9 +227,9 @@ class UpdateController extends Controller
             ]);
         }
         // version 3.1
-        $free_delivery_over_status= BusinessSetting::where('key','free_delivery_over_status')->first();
-        $free_delivery_over= BusinessSetting::where('key','free_delivery_over')->first()?->value;
-        if($free_delivery_over_status?->value == 1 && $free_delivery_over > 0){
+        $free_delivery_over_status = BusinessSetting::where('key', 'free_delivery_over_status')->first();
+        $free_delivery_over = BusinessSetting::where('key', 'free_delivery_over')->first()?->value;
+        if ($free_delivery_over_status?->value == 1 && $free_delivery_over > 0) {
             $free_delivery_over_status->key = 'admin_free_delivery_status';
             $free_delivery_over_status->save();
             Helpers::businessUpdateOrInsert(['key' => 'admin_free_delivery_option'], [
@@ -241,12 +238,13 @@ class UpdateController extends Controller
         }
 
         $data = DataSetting::where('type', 'login_admin')->pluck('value')->first();
-        return redirect('/login/'.$data);
+        return redirect('/login/' . $data);
     }
 
-    private function set_data(){
-        try{
-            $gateway= ['ssl_commerz_payment',
+    private function set_data()
+    {
+        try {
+            $gateway = ['ssl_commerz_payment',
                 'razor_pay',
                 'paypal',
                 'stripe',
@@ -258,21 +256,21 @@ class UpdateController extends Controller
                 'liqpay',
                 'paytm',
                 'bkash',
-                'paytabs' ];
+                'paytabs'];
 
-            $data= BusinessSetting::whereIn('key',$gateway)->pluck('value','key')->toArray();
+            $data = BusinessSetting::whereIn('key', $gateway)->pluck('value', 'key')->toArray();
 
 
-            foreach($data as $key => $value){
+            foreach ($data as $key => $value) {
 
-                $gateway=$key;
-                if($key == 'ssl_commerz_payment' ){
-                    $gateway='ssl_commerz';
+                $gateway = $key;
+                if ($key == 'ssl_commerz_payment') {
+                    $gateway = 'ssl_commerz';
                 }
 
-                $decoded_value= json_decode($value , true);
-                $data= ['gateway' => $gateway ,
-                    'mode' =>  isset($decoded_value['status']) == 1  ?  'live': 'test'
+                $decoded_value = json_decode($value, true);
+                $data = ['gateway' => $gateway,
+                    'mode' => isset($decoded_value['status']) == 1 ? 'live' : 'test'
                 ];
 
                 if ($gateway == 'ssl_commerz') {
@@ -365,9 +363,9 @@ class UpdateController extends Controller
                     ];
                 }
 
-                $credentials= json_encode(array_merge($data, $additional_data));
+                $credentials = json_encode(array_merge($data, $additional_data));
 
-                $payment_additional_data=['gateway_title' => ucfirst(str_replace('_',' ',$gateway)),
+                $payment_additional_data = ['gateway_title' => ucfirst(str_replace('_', ' ', $gateway)),
                     'gateway_image' => null];
 
                 DB::table('addon_settings')->updateOrInsert(['key_name' => $gateway, 'settings_type' => 'payment_config'], [
@@ -375,8 +373,8 @@ class UpdateController extends Controller
                     'live_values' => $credentials,
                     'test_values' => $credentials,
                     'settings_type' => 'payment_config',
-                    'mode' => isset($decoded_value['status']) == 1  ?  'live': 'test',
-                    'is_active' => isset($decoded_value['status']) == 1  ?  1: 0 ,
+                    'mode' => isset($decoded_value['status']) == 1 ? 'live' : 'test',
+                    'is_active' => isset($decoded_value['status']) == 1 ? 1 : 0,
                     'additional_data' => json_encode($payment_additional_data),
                 ]);
             }
@@ -387,64 +385,65 @@ class UpdateController extends Controller
         return true;
     }
 
-    private function set_sms_data(){
-        try{
-            $sms_gateway= ['twilio_sms',
+    private function set_sms_data()
+    {
+        try {
+            $sms_gateway = ['twilio_sms',
                 'nexmo_sms',
                 'msg91_sms',
                 '2factor_sms'];
 
-            $data= BusinessSetting::whereIn('key',$sms_gateway)->pluck('value','key')->toArray();
-            foreach($data as $key => $value){
-                $decoded_value= json_decode($value , true);
+            $data = BusinessSetting::whereIn('key', $sms_gateway)->pluck('value', 'key')->toArray();
+            foreach ($data as $key => $value) {
+                $decoded_value = json_decode($value, true);
 
                 if ($key == 'twilio_sms') {
-                    $sms_gateway='twilio';
+                    $sms_gateway = 'twilio';
                     $additional_data = [
-                        'status' => data_get($decoded_value,'status',null),
-                        'sid' => data_get($decoded_value,'sid',null),
-                        'messaging_service_sid' =>  data_get($decoded_value,'messaging_service_id',null),
-                        'token' => data_get($decoded_value,'token',null),
-                        'from' =>data_get($decoded_value,'from',null),
-                        'otp_template' => data_get($decoded_value,'otp_template',null),
+                        'status' => data_get($decoded_value, 'status', null),
+                        'sid' => data_get($decoded_value, 'sid', null),
+                        'messaging_service_sid' => data_get($decoded_value, 'messaging_service_id', null),
+                        'token' => data_get($decoded_value, 'token', null),
+                        'from' => data_get($decoded_value, 'from', null),
+                        'otp_template' => data_get($decoded_value, 'otp_template', null),
                     ];
                 } elseif ($key == 'nexmo_sms') {
-                    $sms_gateway='nexmo';
+                    $sms_gateway = 'nexmo';
                     $additional_data = [
-                        'status' => data_get($decoded_value,'status',null),
-                        'api_key' => data_get($decoded_value,'api_key',null),
-                        'api_secret' =>  data_get($decoded_value,'api_secret',null),
-                        'token' => data_get($decoded_value,'token',null),
-                        'from' =>  data_get($decoded_value,'from',null),
-                        'otp_template' =>  data_get($decoded_value,'otp_template',null),
+                        'status' => data_get($decoded_value, 'status', null),
+                        'api_key' => data_get($decoded_value, 'api_key', null),
+                        'api_secret' => data_get($decoded_value, 'api_secret', null),
+                        'token' => data_get($decoded_value, 'token', null),
+                        'from' => data_get($decoded_value, 'from', null),
+                        'otp_template' => data_get($decoded_value, 'otp_template', null),
                     ];
                 } elseif ($key == '2factor_sms') {
-                    $sms_gateway='2factor';
+                    $sms_gateway = '2factor';
                     $additional_data = [
-                        'status' => data_get($decoded_value,'status',null),
-                        'api_key' => data_get($decoded_value,'api_key',null),
-                        'otp_template' => data_get($decoded_value,'otp_template','Your OTP is: #OTP#'),
+                        'status' => data_get($decoded_value, 'status', null),
+                        'api_key' => data_get($decoded_value, 'api_key', null),
+                        'otp_template' => data_get($decoded_value, 'otp_template', 'Your OTP is: #OTP#'),
                     ];
                 } elseif ($key == 'msg91_sms') {
-                    $sms_gateway='msg91';
+                    $sms_gateway = 'msg91';
                     $additional_data = [
-                        'status' => data_get($decoded_value,'status',null),
-                        'template_id' =>  data_get($decoded_value,'template_id',null),
-                        'auth_key' =>  data_get($decoded_value,'authkey',null),
+                        'status' => data_get($decoded_value, 'status', null),
+                        'template_id' => data_get($decoded_value, 'template_id', null),
+                        'auth_key' => data_get($decoded_value, 'authkey', null),
                     ];
                 }
-                $data= ['gateway' => $sms_gateway ,
-                    'mode' =>  isset($decoded_value['status']) == 1  ?  'live': 'test'
+                $data = ['gateway' => $sms_gateway,
+                    'mode' => isset($decoded_value['status']) == 1 ? 'live' : 'test'
                 ];
-                $credentials= json_encode(array_merge($data, $additional_data));
+                $credentials = json_encode(array_merge($data, $additional_data));
 
                 DB::table('addon_settings')->updateOrInsert(['key_name' => $sms_gateway, 'settings_type' => 'sms_config'], [
                     'key_name' => $sms_gateway,
                     'live_values' => $credentials,
                     'test_values' => $credentials,
                     'settings_type' => 'sms_config',
-                    'mode' => isset($decoded_value['status']) == 1  ?  'live': 'test',
-                    'is_active' => isset($decoded_value['status']) == 1  ?  1: 0 ,
+                    'mode' => isset($decoded_value['status']) == 1 ? 'live' : 'test',
+                    'is_active' => isset($decoded_value['status']) == 1 ? 1 : 0,
                 ]);
             }
         } catch (\Exception $exception) {
@@ -455,7 +454,8 @@ class UpdateController extends Controller
     }
 
 
-    private function update_table(){
+    private function update_table()
+    {
 
 
         $gateways = [
@@ -546,5 +546,6 @@ class UpdateController extends Controller
             //
         }
     }
+
 
 }
