@@ -1,61 +1,111 @@
 <?php
-ini_set('memory_limit',-1);
-/*
-|--------------------------------------------------------------------------
-| Create The Application
-|--------------------------------------------------------------------------
-|
-| The first thing we will do is create a new Laravel application instance
-| which serves as the "glue" for all the components of Laravel, and is
-| the IoC container for the system binding all of the various parts.
-|
-*/
 
-$app = new Illuminate\Foundation\Application(
-    $_ENV['APP_BASE_PATH'] ?? dirname(__DIR__)
-);
+use App\Http\Middleware\ActivationCheckMiddleware;
+use App\Http\Middleware\AdminMiddleware;
+use App\Http\Middleware\AdminRentalModuleCheckMiddleware;
+// Core Laravel web middleware
+use App\Http\Middleware\APIGuestMiddleware;
+use App\Http\Middleware\Authenticate;
+use App\Http\Middleware\CurrentModule;
+use App\Http\Middleware\DmTokenIsValid;
+use App\Http\Middleware\EncryptCookies;
+use App\Http\Middleware\InstallationMiddleware;
+use App\Http\Middleware\Localization;
+// Custom middleware
+use App\Http\Middleware\LocalizationMiddleware;
+use App\Http\Middleware\ModuleCheckMiddleware;
+use App\Http\Middleware\ModulePermissionMiddleware;
+use App\Http\Middleware\ProviderRentalModuleCheckMiddleware;
+use App\Http\Middleware\ReactValid;
+use App\Http\Middleware\RedirectIfAuthenticated;
+use App\Http\Middleware\Subscription;
+use App\Http\Middleware\VendorMiddleware;
+use App\Http\Middleware\VendorTokenIsValid;
+use App\Http\Middleware\VerifyCsrfToken;
+use Illuminate\Auth\Middleware\AuthenticateWithBasicAuth;
+use Illuminate\Auth\Middleware\Authorize;
+use Illuminate\Auth\Middleware\EnsureEmailIsVerified;
+use Illuminate\Auth\Middleware\RequirePassword;
+use Illuminate\Cookie\Middleware\AddQueuedCookiesToResponse;
+use Illuminate\Foundation\Application;
+use Illuminate\Foundation\Configuration\Exceptions;
+use Illuminate\Foundation\Configuration\Middleware;
+use Illuminate\Http\Middleware\SetCacheHeaders;
+use Illuminate\Routing\Middleware\SubstituteBindings;
+use Illuminate\Routing\Middleware\ThrottleRequests;
+use Illuminate\Routing\Middleware\ValidateSignature;
+use Illuminate\Session\Middleware\StartSession;
+use Illuminate\View\Middleware\ShareErrorsFromSession;
 
-/*
-|--------------------------------------------------------------------------
-| Bind Important Interfaces
-|--------------------------------------------------------------------------
-|
-| Next, we need to bind some important interfaces into the container so
-| we will be able to resolve them when needed. The kernels serve the
-| incoming requests to this application from both the web and CLI.
-|
-*/
+return Application::configure(basePath: dirname(__DIR__))
 
-$app->singleton(
-    Illuminate\Contracts\Http\Kernel::class,
-    App\Http\Kernel::class
-);
+    ->withRouting(
+        // commands: __DIR__ . '/../routes/console.php',
+        health: '/up',
+    )
 
-$app->singleton(
-    Illuminate\Contracts\Console\Kernel::class,
-    App\Console\Kernel::class
-);
+    ->withMiddleware(function (Middleware $middleware) {
 
-$app->singleton(
-    Illuminate\Contracts\Debug\ExceptionHandler::class,
-    App\Exceptions\Handler::class
-);
+        $middleware->use([
+            \App\Http\Middleware\PreventRequestsDuringMaintenance::class,
+            \Illuminate\Foundation\Http\Middleware\ValidatePostSize::class,
+            \App\Http\Middleware\TrimStrings::class,
+            \Illuminate\Foundation\Http\Middleware\ConvertEmptyStringsToNull::class,
+            \Illuminate\Http\Middleware\HandleCors::class,
+        ]);
 
-/*
-|--------------------------------------------------------------------------
-| Return The Application
-|--------------------------------------------------------------------------
-|
-| This script returns the application instance. The instance is given to
-| the calling script so we can separate the building of the instances
-| from the actual running of the application and sending responses.
-|
-*/
+        $middleware->group('web', [
+            EncryptCookies::class,
+            AddQueuedCookiesToResponse::class,
+            StartSession::class,
+            ShareErrorsFromSession::class,
+            VerifyCsrfToken::class,
+            SubstituteBindings::class,
+            Localization::class,
+        ]);
+
+        $middleware->group('api', [
+            SubstituteBindings::class,
+        ]);
+
+        $middleware->alias([
+            'auth' => Authenticate::class,
+            'guest' => RedirectIfAuthenticated::class,
+
+            'admin' => AdminMiddleware::class,
+            'vendor' => VendorMiddleware::class,
+            'vendor.api' => VendorTokenIsValid::class,
+            'dm.api' => DmTokenIsValid::class,
+            'module' => ModulePermissionMiddleware::class,
+            'installation-check' => InstallationMiddleware::class,
+            'actch' => ActivationCheckMiddleware::class,
+            'localization' => LocalizationMiddleware::class,
+            'subscription' => Subscription::class,
+            'react' => ReactValid::class,
+            'apiGuestCheck' => APIGuestMiddleware::class,
+            'auth.basic' => AuthenticateWithBasicAuth::class,
+            'cache.headers' => SetCacheHeaders::class,
+            'can' => Authorize::class,
+            'password.confirm' => RequirePassword::class,
+            'signed' => ValidateSignature::class,
+            'throttle' => ThrottleRequests::class,
+            'verified' => EnsureEmailIsVerified::class,
+            'module-check' => ModuleCheckMiddleware::class,
+            'current-module' => CurrentModule::class,
+            'admin-rental-module' => AdminRentalModuleCheckMiddleware::class,
+            'provider-rental-module' => ProviderRentalModuleCheckMiddleware::class,
+        ]);
+    })
+
+    ->withExceptions(function (Exceptions $exceptions) {
+        //
+    })
+
+    ->create();
+
 // $requestUri = $_SERVER['REQUEST_URI'] ?? '';
-// if (!str_starts_with($requestUri, '/image-proxy')){
+// if (!str_starts_with($requestUri, '/image-proxy')) {
 //     header('Access-Control-Allow-Origin: *');
 // }
-//header('Access-Control-Allow-Methods: *');
-//header('Access-Control-Allow-Headers: *');
-
-return $app;
+// header('Access-Control-Allow-Methods: *');
+// header('Access-Control-Allow-Headers: *');
