@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\Vendor\ItemController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\Vendor\SubscriptionController;
 
@@ -151,9 +152,9 @@ Route::group(['namespace' => 'Vendor', 'as' => 'vendor.'], function () {
             Route::post('bulk-import', 'ItemController@bulk_import_data');
             Route::get('bulk-export', 'ItemController@bulk_export_index')->name('bulk-export-index');
             Route::post('bulk-export', 'ItemController@bulk_export_data')->name('bulk-export');
-
-
             Route::get('flash-sale', 'ItemController@flash_sale')->name('flash_sale');
+
+             Route::get('get-brand-list', [ItemController::class, 'getBrandList'])->name('getBrandList');
 
         });
 
