@@ -151,12 +151,11 @@ if (!$is_published) {
         //PAYSTACK
         Route::group(['prefix' => 'paystack', 'as' => 'paystack.'], function () {
             Route::get('pay', [PaystackController::class, 'index'])->name('pay');
-            Route::post('payment', [PaystackController::class, 'redirectToGateway'])->name('payment');
             Route::get('callback', [PaystackController::class, 'handleGatewayCallback'])->name('callback');
+            Route::get('cancel', [PaystackController::class, 'cancel'])->name('cancel');
         });
 
         //BKASH
-
         Route::group(['prefix' => 'bkash', 'as' => 'bkash.'], function () {
             // Payment Routes for bKash
             Route::get('make-payment', [BkashPaymentController::class, 'make_tokenize_payment'])->name('make-payment');
@@ -214,6 +213,7 @@ Route::group(['prefix' => 'vendor', 'as' => 'restaurant.'], function () {
 
     Route::get('back', 'VendorController@back')->name('back');
     Route::post('business-plan', 'VendorController@business_plan')->name('business_plan');
+    Route::get('business-plan', 'VendorController@secondStep')->name('secondStep');
     Route::post('payment', 'VendorController@payment')->name('payment');
     Route::get('final-step', 'VendorController@final_step')->name('final_step');
 });
